==> results/05_grp34_analysis.Rmd <==
---
title: "Group 3 and 4 analysis"
author: "Kent Riemondy RBI"
date: "`r Sys.Date()`"
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "html") })
output:
  html_document:
    toc: true
    toc_float: true
    toc_collapsed: false
    theme: cosmo
    highlight: tango
    fig_caption: true
    code_folding: hide
    df_print: paged
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r libs, message=FALSE, warning=FALSE, echo=FALSE}
source(here::here("R/utils.R"))

rmd_dir <- "gp34"

fig_dir <- file.path(fig_dir, rmd_dir)
mkrs_dir <- file.path(mkrs_dir, rmd_dir)
tbls_dir <- file.path(tbls_dir, rmd_dir)
obj_dir <- file.path(obj_dir, rmd_dir)

walk(c(fig_dir, mkrs_dir, tbls_dir, obj_dir),
     dir.create, showWarnings = F)


seed_value <- 20200204
```

```{r}
so <- qread(file.path("objects", "preprocess", "all_neoplastic.qs"),
            nthreads = 4)
```


```{r}
gp_cells <- colnames(so)[!so$subgroup %in% c("SHH", "WNT")]

so <- so[, gp_cells]

so <- FindVariableFeatures(
  so,
  selection.method = "vst",
  nfeatures = 3000,
  verbose = FALSE
)

so <- ScaleData(so,
                features = VariableFeatures(so), 
                vars.to.regress = "percent.mt", 
                verbose = TRUE)

so <- RunPCA(so, 
             features = VariableFeatures(so),
             npcs = 50, 
             verbose = FALSE,
             seed.use = seed_value)

ElbowPlot(so, ndims = 50)

# make graphs and use graphs for UMAP
so <- FindNeighbors(so, 
                      reduction = "pca", 
                      dims = 1:50, 
                      k.param = 30L)


so <- RunUMAP(so, 
              dims = 1:50, 
              n.neighbors = 30L,
              min.dist = 0.4,
              seed.use = seed_value)

so <- FindClusters(so, 
                   resolution = c(0.1, 0.3, 0.5), 
                   random.seed = seed_value)

so$coarse_clusters_gp34 <- so$RNA_snn_res.0.1

Idents(so) <- "coarse_clusters_gp34"

plot_umap(so, "coarse_clusters_gp34")
plot_umap(so, "UPN")
plot_umap(so, "subtype")
plot_umap(so, "subgroup")
```

## Harmonize

```{r}
library(harmony)
set.seed(seed_value)
so <- RunHarmony(so,
                 "UPN", 
                 theta = 1.5,
                 plot_convergence = TRUE)

so <- RunUMAP(so, 
              reduction = "harmony", 
              dims = 1:50, 
              min.dist = 0.3,
              n.neighbors = 30L,
              seed.use = seed_value,
              reduction.name = "harmony_umap",
              reduction.key = "UMAP_Harmony_")

so <- FindNeighbors(so,
                    k.param = 20L,
                    reduction = "harmony", 
                    dims = 1:50)

so <- FindClusters(so, 
                   resolution = 0.25, 
                   random.seed = seed_value)

so$coarse_clusters_gp34_harmony <- so$seurat_clusters
Idents(so) <- "coarse_clusters_gp34_harmony"
plot_harmony(so, "coarse_clusters_gp34_harmony", sorted = "random")
```

```{r markers of each subgroup}
full_mkrs <- wilcoxauc(so, "coarse_clusters_gp34_harmony")

mkrs <- filter(full_mkrs, logFC > 0, padj < 0.05, pct_in > 0.10) %>% 
  group_by(group) %>% 
  arrange(padj, desc(logFC), .by_group = TRUE)

mkrs %>% 
  write_tsv(file.path(mkrs_dir, "harmony_markers_gp34.tsv"))

mkrs %>% 
  ungroup() %>% 
  split(., .$group) %>% 
  write_markers_xlsx(.,
                     file.path(tbls_dir,
                               "harmony_markers_gp34.xlsx"))
```

## GO terms

```{r}
library(gprofiler2)
library(ggrepel)
mkrs <- read_tsv(file.path(mkrs_dir, 
                    "harmony_markers_gp34.tsv")) %>% 
  filter( padj < 0.05,
          pct_in > 0.10) %>% 
  group_by(group) %>% 
  slice(1:500)

mkrs_split <- split(mkrs, mkrs$group) %>% 
  map(~pull(.x, feature))

go_res <- gost(mkrs_split, 
               ordered_query = TRUE,
               sources = c("GO", 
                           "KEGG", 
                           "REAC", 
                           "TF"))

go_res[["result"]] %>% 
  mutate(parents = unlist(map(parents, 
                              ~str_c(.x, collapse = ",")))) %>% 
  write_tsv(file.path(tbls_dir, "goterms_gp34_clusters.tsv"))

go_res[["result"]] %>% 
  mutate(parents = unlist(map(parents, 
                              ~str_c(.x, collapse = ","))),
         source = str_replace_all(source, 
                                  "[[:punct:]]", " ")) %>% 
  select(-significant) %>% 
  split(., .$query) %>% 
  openxlsx::write.xlsx(., file.path(tbls_dir,
                                    "goterms_gp34_cluster.xlsx"))
```

## Heatmap of gene expression

```{r}
topx <- read_tsv(file.path(mkrs_dir, 
                    "harmony_markers_gp34.tsv")) %>% 
  filter(padj < 0.05) %>% 
   mutate(group = as.factor(as.integer(group))) %>% 
  group_by(group) %>% 
  arrange(padj, desc(logFC), .by_group = TRUE) %>% 
  slice(1:10)

so <- ScaleData(so, features = rownames(so))

tmp_so <- subset(so, 
                 subset = coarse_clusters_gp34_harmony %in% as.character(c(0:5)))

tmp_so$coarse_clusters_gp34_harmony <- factor(tmp_so$coarse_clusters_gp34_harmony)

p <- DoHeatmap(tmp_so, 
               group.colors = discrete_palette_default,
               features = unique(topx$feature),
               group.by = "coarse_clusters_gp34_harmony",
               angle = 0, 
               raster = FALSE, 
               draw.lines = TRUE)

p <- p +
  scale_fill_gradientn(colours = viridis::viridis(256),
                       name = "Expression\nZ-scores")

save_plot(file.path(fig_dir, "heatmap_markers_gp34.pdf"), 
          p,
          base_width = 10,
          base_height = 12)

rm(tmp_so)
gc()
```


## Add in Northcutt paper metamodule scores

```{r}
supp_tbl_2 <- "https://static-content.springer.com/esm/art%3A10.1038%2Fs41586-019-1434-6/MediaObjects/41586_2019_1434_MOESM3_ESM.xlsx"

if(!file.exists("../docs/northcutt_sup_tbl_2.xlsx")){
  download.file(supp_tbl_2, "../docs/northcutt_sup_tbl_2.xlsx")
}

sup_tbl <- read_excel("../docs/northcutt_sup_tbl_2.xlsx", 
           sheet = 2,
           skip = 3) 

sup_tbl[, 1] <- NULL

# extract out each metaprogram following their approach of excluding ribosomal
# and protein genes and non-coding RNAs

mps <- map(seq(1, ncol(sup_tbl), 2), 
    ~c(.x, .x + 1)) %>% 
  map(., ~sup_tbl[, .x] 
      %>% na.omit()) 

mp_names <- map_chr(mps, ~colnames(.x)[2])
mps <- map(mps, ~.x[, 2, drop = TRUE]) 
names(mps) <- mp_names

gp34_mps <- mps[str_subset(names(mps), "Group")]

mp_ids <- c(
  "Group 3/4-A-cell_cycle",
  "Group 3/4-B-Translation_Myc",
  "Group 3/4-C-Differentiation"
)

names(gp34_mps) <- mp_ids

for (i in seq_along(gp34_mps)){
  so <- AddModuleScore(so, 
                       features = list(c(gp34_mps[[i]])),
                       ctrl = 50,
                       name = names(gp34_mps)[i],
                       seed = 42)
}

new_ids <- str_c(make.names(mp_ids), "1") 

new_id_idx <- match(new_ids, colnames(so@meta.data))
colnames(so@meta.data)[new_id_idx] <- mp_ids
```


```{r}
so <- CellCycleScoring(so,
                       g2m.features = cc.genes$g2m.genes,
                       s.features = cc.genes$s.genes)
```


## Write out matrix for other tools

```{r}
#full count matrix
GetAssayData(so, "counts") %>% 
  as.matrix(.) %>% 
  as.data.frame() %>% 
  rownames_to_column("gene") %>% 
  data.table::fwrite(., file.path(tbls_dir, "gp34_count_matrix.tsv.gz"),
       sep = "\t",
       compress = "gzip")
  
# expression matrix filtered at least 25 cells express the gene
expr_mat <- GetAssayData(so, "data")
expr_mat <- expr_mat[Matrix::rowSums(expr_mat > 0) >= 25, ]

expr_mat %>% 
  as.matrix(.) %>% 
  as.data.frame() %>% 
  rownames_to_column("gene") %>% 
  data.table::fwrite(., file.path(tbls_dir, "gp34_expr_matrix.tsv.gz"),
       sep = "\t",
       compress = "gzip")

rm(expr_mat); gc()

```

save loom file for paga

```{r, eval = FALSE}

# drop graphs see https://github.com/theislab/scanpy/issues/598
tmp_obj <- qread("objects/gp34.qs")
tmp_obj@graphs <- list()
tmp_obj@meta.data <- tmp_obj@meta.data[, c("coarse_clusters_gp34_harmony",
                                           "orig.ident",
                                           "subtype",
                                           "subgroup",
                                           "UPN")]

lfile <- as.loom(tmp_obj,filename = file.path("objects", "grp34.loom"),
        verbose = TRUE,
        overwrite = TRUE)
lfile$close_all()

rm(tmp_obj)
```

## Get top PC genes

```{r}
to_plot <- c("UPN", "coarse_cell_type", "subgroup")

dims_to_plot <- list(
  c(1,2),
  c(2,3),
  c(3,4),
  c(4,5))
  
plts <- list()
for ( i in seq_along(dims_to_plot)){
  plts[[i]] <- map(to_plot, ~plot_pca(so, .x, dims =dims_to_plot[[i]])) %>%
  plot_grid(plotlist = ., nrow = 1, ncol = length(to_plot),
            rel_widths = c(1.75, 1.5, 1))
}

names(plts) <- str_c("pc_", 
                     map_chr(dims_to_plot, ~str_c(.x, collapse = "_")),
                     ".pdf")
pc_dir <- file.path(fig_dir, "pcs_grp34")
dir.create(pc_dir)
iwalk(plts,
  ~save_plot(file.path(pc_dir, .y), .x, nrow = 1, ncol = 3))


pc_loadings <- so@reductions$pca@feature.loadings[, 1:20] %>% 
  as.data.frame() %>% 
  tibble::rownames_to_column("gene") %>% 
  arrange(desc(PC_1))


pc_tbl_dir <- file.path(tbls_dir, "pcs", "grp34")
dir.create(pc_tbl_dir, recursive = TRUE)
write_tsv(pc_loadings, file.path(pc_tbl_dir,
                                 "pc_loadings.tsv.gz"))

tibble(pc = str_c("PC", 1:length(so@reductions$pca@stdev)), 
                  stdev = so@reductions$pca@stdev) %>% 
  write_tsv(file.path(pc_tbl_dir, "pc_variance_explained.tsv"))
```



## Average Expression per cluster

```{r}
plot_harmony(so, 
             "coarse_clusters_gp34_harmony")


write_avg_expr(so,  
               "coarse_clusters_gp34_harmony",
               file.path(tbls_dir, 
                         "coarse_clusters_gp34_harmony_avgexpr.tsv.gz"))


cell_count_mat <- get_cell_count_matrix(so, "UPN", "coarse_clusters_gp34_harmony")

cell_count_mat[is.na(cell_count_mat)] <- 0L

cell_count_mat %>% 
  tibble::rownames_to_column("UPN") %>% 
write_csv(file.path(tbls_dir, 
                         "coarse_clusters_gp34_harmony_cell_counts.csv"))


full_marker_matrix <- full_mkrs %>% 
  top_marker_matrix(.) 
  
topn_marker_matrix <- full_mkrs %>% 
  top_marker_matrix(n = 200) 

out <- list(avg_expression = read_tsv(file.path(tbls_dir, "coarse_clusters_gp34_harmony_avgexpr.tsv.gz")),
     logFC_all = full_marker_matrix,
     top_200_logFC = topn_marker_matrix)

openxlsx::write.xlsx(out, 
                     file.path(tbls_dir, "coarse_clusters_gp34_harmony_expression_logfc_summaries.xlsx"))
```


## CytoTRACE

```{r}
source("../R/CytoTRACE.R")
#subset to genes expressed in at least 25 cells
to_keep <- Matrix::rowSums(so@assays$RNA@counts > 0) > 25

a <- CytoTRACE(as.matrix(so@assays$RNA@counts[to_keep, ]),
               ncores = 4)

res <- tibble(
  cell = names(a$GCS),
  GCS = a$GCS,
  CytoTRACE = a$CytoTRACE) %>% 
  left_join(get_metadata(so), ., by = "cell") %>% 
  select(cell, GCS, CytoTRACE) %>% 
  mutate(GCS = ifelse(is.na(GCS), 0, GCS),
         CytoTRACE = ifelse(is.na(CytoTRACE), -1, CytoTRACE)) %>% 
  column_to_rownames("cell")

so <- AddMetaData(so, res)

plot_harmony(so, 
             "CytoTRACE",
             sorted = FALSE)

plot_harmony(so, 
             "GCS",
             sorted = FALSE)

plot_harmony(so, 
             "nFeature_RNA",
             sorted = FALSE)
```

## cluster diversity

```{r}
a <- calc_diversity(so, "UPN", "coarse_clusters_gp34")@meta.data %>% 
  select(unaligned_cluster_diversity = entropy)

b <- calc_diversity(so, "UPN", "coarse_clusters_gp34_harmony")@meta.data %>% 
  select(aligned_cluster_diversity = entropy)

so <- AddMetaData(so, cbind(a, b))
```

## Save 

```{r}
qsave(so, file.path(obj_dir, "gp34.qs"), nthreads = 4)
```

## Group3 only analysis

```{r}
so_gp3 <- subset(so, subset = subgroup %in% c("GP3", "GP3/4"))
so_gp4 <- subset(so, subset = subgroup == "GP4")

rm(so); gc()

gp3_tbls_dir <- file.path(proj_dir, "results", "tables", "gp3")
dir.create(gp3_tbls_dir, showWarnings = FALSE)
```


```{r}
so <- FindVariableFeatures(
  so_gp3,
  selection.method = "vst",
  nfeatures = 3000,
  verbose = FALSE
)

so <- ScaleData(so,
                features = VariableFeatures(so), 
                verbose = TRUE)

so <- RunPCA(so, 
             features = VariableFeatures(so),
             npcs = 50, 
             seed.use = seed_value,
             verbose = FALSE)

ElbowPlot(so, ndims = 50)

so <- FindNeighbors(so, 
                    reduction = "pca", 
                    dims = 1:30, 
                    k.param = 20L)

so <- RunUMAP(so, 
              dims = 1:50, 
              n.neighbors = 30L,
              seed.use = seed_value,
              min.dist = 0.4)

so <- FindClusters(so, 
                   resolution = c(0.1, 0.3, 0.5),
                   random.seed = seed_value)

so$coarse_clusters_gp3 <- so$RNA_snn_res.0.1

Idents(so) <- "coarse_clusters_gp3"

plot_umap(so, c("coarse_clusters_gp3", "UPN", "subtype"))
```

## Harmonize

```{r}
set.seed(seed_value + 1)
so <- RunHarmony(so,
                 "UPN", 
                 theta = 1.5,
                 plot_convergence = TRUE)

so <- RunUMAP(so, 
              reduction = "harmony", 
              dims = 1:30, 
              min.dist = 0.4,
              n.neighbors = 20L,
              seed.use =  seed_value,
              reduction.name = "harmony_umap",
              reduction.key = "UMAP_Harmony_")

so <- FindNeighbors(so, reduction = "harmony", dims = 1:30)
so <- FindClusters(so, resolution = 0.2, random.seed =  seed_value)

so$coarse_clusters_gp3_harmony <- so$seurat_clusters
Idents(so) <- "coarse_clusters_gp3_harmony"

plot_harmony(so, c("NRL", "UPN","subtype", "coarse_clusters_gp3_harmony"))
```


```{r}
full_mkrs <- wilcoxauc(so, "coarse_clusters_gp3_harmony")

mkrs <- filter(full_mkrs, logFC > 0, padj < 0.05, pct_in > 0.10) %>% 
  group_by(group) %>% 
  arrange(padj, desc(logFC), .by_group = TRUE)

mkrs %>% 
  write_tsv(file.path(mkrs_dir, "harmony_markers_gp3_only.tsv"))

mkrs %>% 
  ungroup() %>% 
  split(., .$group) %>% 
  write_markers_xlsx(.,
                     file.path(gp3_tbls_dir, "harmony_markers_gp3_only.xlsx"))
```


```{r}
#full count matrix
GetAssayData(so, "counts") %>% 
  as.matrix(.) %>% 
  as.data.frame() %>% 
  rownames_to_column("gene") %>% 
  data.table::fwrite(., file.path(gp3_tbls_dir, "gp3_count_matrix.tsv.gz"),
       sep = "\t",
       compress = "gzip")
  
# expression matrix filtered at least 25 cells express the gene
expr_mat <- GetAssayData(so, "data")
expr_mat <- expr_mat[Matrix::rowSums(expr_mat > 0) >= 25, ]

expr_mat %>% 
  as.matrix(.) %>% 
  as.data.frame() %>% 
  rownames_to_column("gene") %>% 
  data.table::fwrite(., file.path(gp3_tbls_dir, "gp3_expr_matrix.tsv.gz"),
       sep = "\t",
       compress = "gzip")



cell_count_mat <- get_cell_count_matrix(so, "UPN", "coarse_clusters_gp3_harmony")

cell_count_mat[is.na(cell_count_mat)] <- 0L

cell_count_mat %>% 
  tibble::rownames_to_column("UPN") %>% 
write_csv(file.path(gp3_tbls_dir, 
                         "coarse_clusters_gp3_harmony_cell_counts.csv"))
```

## CytoTRACE

```{r}
source("../R/CytoTRACE.R")
#subset to genes expressed in at least 25 cells
to_keep <- Matrix::rowSums(so@assays$RNA@counts > 0) > 25

a <- CytoTRACE(as.matrix(so@assays$RNA@counts[to_keep, ]),
               ncores = 4)

mdata <- get_metadata(so, embedding = NULL) %>% 
  select(-GCS, -CytoTRACE)

res <- tibble(
  cell = names(a$GCS),
  GCS = a$GCS,
  CytoTRACE = a$CytoTRACE) %>% 
  left_join(mdata, ., by = "cell") %>% 
  select(cell, GCS, CytoTRACE) %>% 
  mutate(GCS = ifelse(is.na(GCS), 0, GCS),
         CytoTRACE = ifelse(is.na(CytoTRACE), -1, CytoTRACE)) %>% 
  column_to_rownames("cell")

so <- AddMetaData(so, res)

plot_harmony(so, 
             "CytoTRACE",
             sorted = FALSE)

plot_harmony(so, 
             "GCS",
             sorted = FALSE)

plot_harmony(so, 
             "nFeature_RNA",
             sorted = FALSE)
```


## cluster diversity

```{r}
a <- calc_diversity(so, "UPN", "coarse_clusters_gp3")@meta.data %>% 
  select(unaligned_cluster_diversity = entropy)

b <- calc_diversity(so, "UPN", "coarse_clusters_gp3_harmony")@meta.data %>% 
  select(aligned_cluster_diversity = entropy)

so <- AddMetaData(so, cbind(a, b))
```


```{r}
qsave(so, file.path(obj_dir, "gp3_only.qs"), nthreads = 4)
```

```{r, eval = FALSE}
so <- qread(file.path(obj_dir, "gp3_only.qs"), nthreads = 4)
```

## GO terms

```{r}
library(gprofiler2)
library(ggrepel)
mkrs <- read_tsv(file.path(mkrs_dir, 
                    "harmony_markers_gp3_only.tsv")) %>% 
  filter( padj < 0.05,
          pct_in > 0.10) %>% 
  group_by(group) %>% 
  slice(1:500)

mkrs_split <- split(mkrs, mkrs$group) %>% 
  map(~pull(.x, feature))

go_res <- gost(mkrs_split, 
               ordered_query = TRUE,
               sources = c("GO", 
                           "KEGG", 
                           "REAC", 
                           "TF"))

go_res[["result"]] %>% 
  mutate(parents = unlist(map(parents, 
                              ~str_c(.x, collapse = ",")))) %>% 
  write_tsv(file.path(gp3_tbls_dir, "goterms_gp3_clusters.tsv"))

go_res[["result"]] %>% 
  mutate(parents = unlist(map(parents, 
                              ~str_c(.x, collapse = ","))),
         source = str_replace_all(source, 
                                  "[[:punct:]]", " ")) %>% 
  select(-significant) %>% 
  split(., .$query) %>% 
  openxlsx::write.xlsx(., file.path(gp3_tbls_dir,
                                    "goterms_gp3_cluster.xlsx"))
```

## Cell type naming

```{r}

new_ids <- c(
    "0" = "Differentiated (GOLGA8A, MALAT1) #0",
    "1" = "Differentiated Photoreceptor (NRL) #1",
    "3" = "Cell Cycle #3" ,
    "2" = "Undifferentiated (RPGs and BTF3) #2",
    "4" = "Undifferentiated (MYC, FOXQ1/F2) #4"
  )

so$tumor_cell_type_gp3 <- new_ids[as.character(so$coarse_clusters_gp3_harmony)]

# use andy's names
new_ids <- c(
    "0" = "GP3-C1",
    "1" = "GP3-C2",
    "3" = "GP3-A" ,
    "2" = "GP3-B1",
    "4" = "GP3-B2"
  )

so$tumor_subpopulations_gp3 <- new_ids[as.character(so$coarse_clusters_gp3_harmony)]

plot_harmony(so, c("tumor_cell_type_gp3", "tumor_subpopulations_gp3"))
```

Back add the cell classifications to all neoplastic object

```{r}
so_mdata <- get_metadata(so) %>% 
  select(cell, tumor_cell_type_gp3, tumor_subpopulations_gp3)

so_neo <- qread(file.path("objects", "preprocess", "all_neoplastic.qs"))

new_mdata <- get_metadata(so_neo) %>% 
  select(cell,
         tumor_cell_types_all_subgroups, 
         tumor_cell_type_subgroup_specific,
         tumor_subpopulation) %>% 
  left_join(so_mdata, by = "cell") %>% 
  mutate(tumor_cell_type_subgroup_specific = ifelse(!is.na(tumor_cell_type_gp3),
                                                    str_c("GP3_", tumor_cell_type_gp3),
                                                    tumor_cell_type_subgroup_specific),
         tumor_subpopulations_gp3 = ifelse(!is.na(tumor_subpopulations_gp3),
                                      tumor_subpopulations_gp3,
                                      tumor_subpopulation))


so_neo$tumor_cell_type_subgroup_specific <- new_mdata$tumor_cell_type_subgroup_specific
so_neo$tumor_subpopulation <- new_mdata$tumor_subpopulations_gp3
new_mdata <- so_mdata %>% 
  left_join(get_metadata(so_neo), by = "cell") %>% 
  select(cell, tumor_cell_types_all_subgroups)

so$tumor_cell_types_all_subgroups <- new_mdata$tumor_cell_types_all_subgroups
```


## save

```{r}
qsave(so, file.path(obj_dir, "gp3_only.qs"), nthreads = 4)
qsave(so_neo, file.path("objects", "preprocess", "all_neoplastic.qs"))
```

### Generate additional subclusters

```{r}
cl_dir <- "additional_clustering"

sc_fig_dir <- file.path(proj_dir, "results", "figs", "gp3", cl_dir)
sc_mkrs_dir <- file.path(proj_dir, "results", "markers", "GP3", cl_dir)
sc_tbls_dir <- file.path(gp3_tbls_dir, cl_dir)

walk(c(sc_fig_dir, sc_mkrs_dir, sc_tbls_dir),
     dir.create, showWarnings = F)
```

```{r}
seed_value <- 20200515
so <- FindClusters(so, resolution = 0.5, random.seed = seed_value)
cluster_cols <- str_subset(colnames(so@meta.data), "RNA_snn_res")
plot_harmony(so, cluster_cols)

so$refined_clusters_gp3_harmony <- so$RNA_snn_res.0.5

plot_harmony(so, "refined_clusters_gp3_harmony")
```


```{r}
# make labels based on subpoplutation labels

mdata <- get_metadata(so) %>% 
  select(cell, tumor_subpopulations_gp3, refined_clusters_gp3_harmony)

so <- mdata %>% 
  group_by(tumor_subpopulations_gp3, refined_clusters_gp3_harmony, .drop = FALSE) %>% 
  summarize(n = n()) %>% 
  group_by(refined_clusters_gp3_harmony) %>% 
  mutate(cluster_prop = n / sum(n)) %>% 
  select(-n) %>%
  group_by(refined_clusters_gp3_harmony) %>% 
  arrange(desc(cluster_prop)) %>% 
  slice(1) %>% # get subpopulation assignment with highest proportion (all > 0.6)
  group_by(tumor_subpopulations_gp3) %>% 
  arrange(refined_clusters_gp3_harmony, .by_group = TRUE) %>% 
  mutate(additional_subpop_label = as.integer(factor(refined_clusters_gp3_harmony)),
         new_label = str_c(tumor_subpopulations_gp3, "-", additional_subpop_label)) %>% 
  ungroup() %>% 
  select(refined_clusters_gp3_harmony, 
         tumor_subpopulations_gp3_refined = new_label) %>% 
  left_join(mdata, ., by = c("refined_clusters_gp3_harmony")) %>% 
  column_to_rownames("cell") %>% 
  select(tumor_subpopulations_gp3_refined) %>% 
  AddMetaData(so, .)
  
plot_harmony(so, "tumor_subpopulations_gp3_refined")
```


```{r markers of each subgroup}
full_mkrs <- wilcoxauc(so, "tumor_subpopulations_gp3_refined")

full_mkrs %>% 
  write_tsv(file.path(sc_mkrs_dir, "harmony_markers_gp3_all.tsv"))

mkrs <- filter(full_mkrs, logFC > 0, padj < 0.05, pct_in > 0.10) %>% 
  group_by(group) %>% 
  arrange(padj, desc(logFC), .by_group = TRUE)

mkrs %>% 
  write_tsv(file.path(sc_mkrs_dir, "harmony_markers_gp3.tsv"))

mkrs %>% 
  ungroup() %>% 
  split(., .$group) %>% 
  write_markers_xlsx(.,
                   file.path(sc_tbls_dir, "harmony_markers_gp3.xlsx"))
```


```{r}
library(gprofiler2)
library(ggrepel)
# Use top 500 markers 
mkrs <- read_tsv(file.path(sc_mkrs_dir, 
                    "harmony_markers_gp3.tsv")) %>% 
  filter( padj < 0.05) %>% 
  group_by(group) %>% 
  slice(1:500)

mkrs_split <- split(mkrs, mkrs$group) %>% 
  map(~pull(.x, feature))

go_res <- gost(mkrs_split, 
               ordered_query = TRUE,
               sources = c("GO", 
                          "KEGG", 
                          "REAC", 
                          "TF"))

go_res[["result"]] %>% 
  mutate(parents = unlist(map(parents, ~str_c(.x, collapse = ",")))) %>% 
  write_tsv(., file.path(sc_tbls_dir, "goterms_gp3_clusters.tsv"))

go_res[["result"]] %>% 
  mutate(parents = unlist(map(parents, ~str_c(.x, collapse = ","))),
         source = str_replace_all(source, "[[:punct:]]", " ")) %>% 
  select(-significant) %>% 
  split(., .$query) %>% 
  openxlsx::write.xlsx(., file.path(sc_tbls_dir, "goterms_gp3_cluster.xlsx"))
```
 


```{r}
plot_harmony(so, 
             "tumor_subpopulations_gp3_refined")

write_avg_expr(so,  
               "tumor_subpopulations_gp3_refined",
               file.path(sc_tbls_dir, 
                         "tumor_subpopulations_refined_avgexpr.tsv.gz"))


cell_count_mat <- get_cell_count_matrix(so, "UPN", "tumor_subpopulations_gp3_refined")

cell_count_mat[is.na(cell_count_mat)] <- 0L

cell_count_mat %>% 
  tibble::rownames_to_column("UPN") %>% 
write_csv(file.path(sc_tbls_dir, 
                         "tumor_subpopulations_refined_harmony_cell_counts.csv"))

full_marker_matrix <- full_mkrs %>% 
  top_marker_matrix(.) 
  
topn_marker_matrix <- full_mkrs %>% 
  top_marker_matrix(n = 200) 

out <- list(avg_expression = read_tsv(file.path(sc_tbls_dir, "tumor_subpopulations_refined_avgexpr.tsv.gz")),
     logFC_all = full_marker_matrix,
     top_200_logFC = topn_marker_matrix)

openxlsx::write.xlsx(out, 
                     file.path(sc_tbls_dir, "tumor_subpopulations_refined_gp3_harmony_expression_logfc_summaries.xlsx"))
```
## save

```{r}
qsave(so, file.path(obj_dir, "gp3_only.qs"), nthreads = 4)
```

## GP4 only

```{r}
gp4_tbls_dir <- file.path(proj_dir, "results", "tables", "gp4")
dir.create(gp4_tbls_dir, showWarnings = FALSE)
```

```{r}
so <- FindVariableFeatures(
  so_gp4,
  selection.method = "vst",
  nfeatures = 3000,
  verbose = FALSE
)

so <- ScaleData(so, features = VariableFeatures(so), 
                  verbose = TRUE)

so <- RunPCA(so, 
             features = VariableFeatures(so),
             npcs = 50, 
             seed.use = seed_value,
             verbose = FALSE)

ElbowPlot(so, ndims = 50)

# make graphs and use graphs for UMAP
so <- FindNeighbors(so, 
                      reduction = "pca", 
                      dims = 1:30, 
                      k.param = 20L)

so <- RunUMAP(so, 
              dims = 1:50, 
              n.neighbors = 30L,
              seed.use = seed_value,
              min.dist = 0.4)

so <- FindClusters(so, 
                   resolution = c(0.1, 0.3, 0.5),
                   random.seed = seed_value)

so$coarse_clusters_gp4 <- so$RNA_snn_res.0.1

Idents(so) <- "coarse_clusters_gp4"

plot_umap(so, "coarse_clusters_gp4")
plot_umap(so, "UPN")
plot_umap(so, "subtype")
```

## Harmonize

```{r}
set.seed(seed_value)
so <- RunHarmony(so,
                 "UPN", 
                 theta = 1.5,
                 plot_convergence = TRUE)

so <- RunUMAP(so, 
              reduction = "harmony", 
              dims = 1:30, 
              min.dist = 0.3,
              n.neighbors = 20L,
              seed.use =  seed_value,
              reduction.name = "harmony_umap",
              reduction.key = "UMAP_Harmony_")



so <- FindNeighbors(so, reduction = "harmony", dims = 1:30)
so <- FindClusters(so, resolution = 0.2, random.seed =  seed_value)


so$coarse_clusters_gp4_harmony <- so$seurat_clusters
Idents(so) <- "coarse_clusters_gp4_harmony"
plot_harmony(so, "NRL")
plot_harmony(so, "UPN")
plot_harmony(so, "subtype")
plot_harmony(so, "coarse_clusters_gp4_harmony")
```


```{r}
full_mkrs <- wilcoxauc(so, "coarse_clusters_gp4_harmony")

mkrs <- filter(full_mkrs, logFC > 0, padj < 0.05, pct_in > 0.10) %>% 
  group_by(group) %>% 
  arrange(padj, desc(logFC), .by_group = TRUE)

mkrs %>% 
  write_tsv(file.path(mkrs_dir, "harmony_markers_gp4_only.tsv"))

mkrs %>% 
  ungroup() %>% 
  split(., .$group) %>% 
  write_markers_xlsx(.,
                     file.path(gp4_tbls_dir, "harmony_markers_gp4_only.xlsx"))
```

```{r}
#full count matrix
GetAssayData(so, "counts") %>% 
  as.matrix(.) %>% 
  as.data.frame() %>% 
  rownames_to_column("gene") %>% 
  data.table::fwrite(., file.path(gp4_tbls_dir, "gp4_count_matrix.tsv.gz"),
       sep = "\t",
       compress = "gzip")
  
# expression matrix filtered at least 25 cells express the gene
expr_mat <- GetAssayData(so, "data")
expr_mat <- expr_mat[Matrix::rowSums(expr_mat > 0) >= 25, ]

expr_mat %>% 
  as.matrix(.) %>% 
  as.data.frame() %>% 
  rownames_to_column("gene") %>% 
  data.table::fwrite(., file.path(gp4_tbls_dir, "gp4_expr_matrix.tsv.gz"),
       sep = "\t",
       compress = "gzip")

cell_count_mat <- get_cell_count_matrix(so, "UPN", "coarse_clusters_gp4_harmony")

cell_count_mat[is.na(cell_count_mat)] <- 0L

cell_count_mat %>% 
  tibble::rownames_to_column("UPN") %>% 
write_csv( file.path(gp4_tbls_dir, 
                         "coarse_clusters_gp4_harmony_cell_counts.csv"))
```

## CytoTRACE

```{r}
source("../R/CytoTRACE.R")
#subset to genes expressed in at least 25 cells
to_keep <- Matrix::rowSums(so@assays$RNA@counts > 0) > 25

a <- CytoTRACE(as.matrix(so@assays$RNA@counts[to_keep, ]),
               ncores = 4)

mdata <- get_metadata(so, embedding = NULL) %>% 
  select(-GCS, -CytoTRACE)


res <- tibble(
  cell = names(a$GCS),
  GCS = a$GCS,
  CytoTRACE = a$CytoTRACE) %>% 
  left_join(mdata, ., by = "cell") %>% 
  select(cell, GCS, CytoTRACE) %>% 
  mutate(GCS = ifelse(is.na(GCS), 0, GCS),
         CytoTRACE = ifelse(is.na(CytoTRACE), -1, CytoTRACE)) %>% 
  column_to_rownames("cell")

so <- AddMetaData(so, res)

plot_harmony(so, 
             "CytoTRACE",
             sorted = FALSE)

plot_harmony(so, 
             "GCS",
             sorted = FALSE)

plot_harmony(so, 
             "nFeature_RNA",
             sorted = FALSE)
```

## cluster diversity

```{r}
a <- calc_diversity(so, "UPN", "coarse_clusters_gp4")@meta.data %>% 
  select(unaligned_cluster_diversity = entropy)

b <- calc_diversity(so, "UPN", "coarse_clusters_gp4_harmony")@meta.data %>% 
  select(aligned_cluster_diversity = entropy)

so <- AddMetaData(so, cbind(a, b))
```

```{r}
qsave(so, file.path(obj_dir, "gp4_only.qs"), nthreads = 4)
```

## GO terms

```{r}
library(gprofiler2)
library(ggrepel)
mkrs <- read_tsv(file.path(mkrs_dir, 
                    "harmony_markers_gp4_only.tsv")) %>% 
  filter( padj < 0.05,
          pct_in > 0.10) %>% 
  group_by(group) %>% 
  slice(1:500)

mkrs_split <- split(mkrs, mkrs$group) %>% 
  map(~pull(.x, feature))

go_res <- gost(mkrs_split, 
               ordered_query = TRUE,
               sources = c("GO", 
                           "KEGG", 
                           "REAC", 
                           "TF"))

go_res[["result"]] %>% 
  mutate(parents = unlist(map(parents, 
                              ~str_c(.x, collapse = ",")))) %>% 
  write_tsv(file.path(gp4_tbls_dir, "goterms_gp4_clusters.tsv"))

go_res[["result"]] %>% 
  mutate(parents = unlist(map(parents, 
                              ~str_c(.x, collapse = ","))),
         source = str_replace_all(source, 
                                  "[[:punct:]]", " ")) %>% 
  select(-significant) %>% 
  split(., .$query) %>% 
  openxlsx::write.xlsx(., file.path(gp4_tbls_dir,
                                    "goterms_gp4_cluster.xlsx"))
```

## Cell type naming

```{r}

new_ids <- c(
    "0" = "Differentiated (GRIA2, GOLGA8A, MALAT1) #0",
    "5" = "Differentiated Photoreceptor (NRL) #5",
    "4" = "Cell Cycle (EZH2, Differentiated-like) #4",
    "3" = "Cell Cycle (STMN2, Undifferentiated-like) #3",
    "1" = "Undifferentiated (RPGs, STMN1/2, EOMES) #1",
    "2" = "Undifferentiated (RPGs, STMN1/2) #2",
    "6" = "Exclude: Unknown #6",
    "7" = "Exclude: Single tumor #7",
    "8" = "Exclude: Single tumor #8"
  )
so$tumor_cell_type_gp4 <- new_ids[as.character(so$coarse_clusters_gp4_harmony)]

# use andy's names
new_ids <- c(
    "1" = "GP4-B1",
    "2" = "GP4-B2",
    
    "3" = "GP4-A1",
    "4" = "GP4-A2",
    
    "0" = "GP4-C1",
    "5" = "GP4-C2",
    
    "6" = "GP4-X1",
    "7" = "GP4-X2",
    "8" = "GP4-X3"
  )

so$tumor_subpopulations_gp4 <- new_ids[as.character(so$coarse_clusters_gp4_harmony)]

plot_harmony(so, c("tumor_cell_type_gp4", "tumor_subpopulations_gp4"))

```

Back add the cell classifications to all neoplastic object

```{r}
so_mdata <- get_metadata(so) %>% 
  select(cell, tumor_cell_type_gp4, tumor_subpopulations_gp4)

so_neo <- qread(file.path("objects", "preprocess", "all_neoplastic.qs"))

new_mdata <- get_metadata(so_neo) %>% 
  select(cell,
         tumor_cell_types_all_subgroups, 
         tumor_cell_type_subgroup_specific,
         tumor_subpopulation) %>% 
  left_join(so_mdata, by = "cell") %>% 
  mutate(tumor_cell_type_subgroup_specific = ifelse(!is.na(tumor_cell_type_gp4),
                                                    str_c("GP4_", tumor_cell_type_gp4),
                                                    tumor_cell_type_subgroup_specific),
         tumor_subpopulations_gp4 = ifelse(!is.na(tumor_subpopulations_gp4),
                                      tumor_subpopulations_gp4,
                                      tumor_subpopulation))

so_neo$tumor_cell_type_subgroup_specific <- new_mdata$tumor_cell_type_subgroup_specific
so_neo$tumor_subpopulation <- new_mdata$tumor_subpopulations_gp4

new_mdata <- so_mdata %>% 
  left_join(get_metadata(so_neo), by = "cell") %>% 
  select(cell, tumor_cell_types_all_subgroups)

so$tumor_cell_types_all_subgroups <- new_mdata$tumor_cell_types_all_subgroups
```


### Generate additional subclusters

```{r}
cl_dir <- "additional_clustering"

sc_fig_dir <- file.path(proj_dir, "results", "figs", "gp4", cl_dir)
sc_mkrs_dir <- file.path(proj_dir, "results", "markers", "GP4", cl_dir)
sc_tbls_dir <- file.path(gp4_tbls_dir, cl_dir)

walk(c(sc_fig_dir, sc_mkrs_dir, sc_tbls_dir),
     dir.create, showWarnings = F)
```

```{r}
seed_value <- 20200515
so <- FindNeighbors(so, reduction = "harmony", dims = 1:30) %>% 
  FindClusters(resolution =  0.45, random.seed = seed_value)
cluster_cols <- str_subset(colnames(so@meta.data), "RNA_snn_res")
plot_harmony(so, cluster_cols)

so$refined_clusters_gp4_harmony <- so$RNA_snn_res.0.45

plot_harmony(so, "refined_clusters_gp4_harmony")
```


```{r}
# make labels based on subpoplutation labels

mdata <- get_metadata(so) %>% 
  select(cell, tumor_subpopulations_gp4, refined_clusters_gp4_harmony)

so <- mdata %>% 
  group_by(tumor_subpopulations_gp4, refined_clusters_gp4_harmony, .drop = FALSE) %>% 
  summarize(n = n()) %>% 
  group_by(refined_clusters_gp4_harmony) %>% 
  mutate(cluster_prop = n / sum(n)) %>% 
  select(-n) %>%
  group_by(refined_clusters_gp4_harmony) %>% 
  arrange(desc(cluster_prop)) %>% 
  slice(1) %>% 
  group_by(tumor_subpopulations_gp4) %>% 
  arrange(refined_clusters_gp4_harmony, .by_group = TRUE) %>% 
  mutate(additional_subpop_label = as.integer(factor(refined_clusters_gp4_harmony)),
         new_label = str_c(tumor_subpopulations_gp4, "-", additional_subpop_label)) %>% 
  ungroup() %>% 
  select(refined_clusters_gp4_harmony, 
         tumor_subpopulations_gp4_refined = new_label) %>% 
  left_join(mdata, ., by = c("refined_clusters_gp4_harmony")) %>% 
  column_to_rownames("cell") %>% 
  select(tumor_subpopulations_gp4_refined) %>% 
  AddMetaData(so, .)
  
plot_harmony(so, "tumor_subpopulations_gp4_refined")
```


```{r markers of each subgroup}
full_mkrs <- wilcoxauc(so, "tumor_subpopulations_gp4_refined")

full_mkrs %>% 
  write_tsv(file.path(sc_mkrs_dir, "harmony_markers_gp4_all.tsv"))

mkrs <- filter(full_mkrs, logFC > 0, padj < 0.05, pct_in > 0.10) %>% 
  group_by(group) %>% 
  arrange(padj, desc(logFC), .by_group = TRUE)

mkrs %>% 
  write_tsv(file.path(sc_mkrs_dir, "harmony_markers_gp4.tsv"))

mkrs %>% 
  ungroup() %>% 
  split(., .$group) %>% 
  write_markers_xlsx(.,
                   file.path(sc_tbls_dir, "harmony_markers_gp4.xlsx"))
```


```{r}
library(gprofiler2)
library(ggrepel)
# Use top 500 markers 
mkrs <- read_tsv(file.path(sc_mkrs_dir, 
                    "harmony_markers_gp4.tsv")) %>% 
  filter( padj < 0.05) %>% 
  group_by(group) %>% 
  slice(1:500)

mkrs_split <- split(mkrs, mkrs$group) %>% 
  map(~pull(.x, feature))

go_res <- gost(mkrs_split, 
               ordered_query = TRUE,
               sources = c("GO", 
                          "KEGG", 
                          "REAC", 
                          "TF"))

go_res[["result"]] %>% 
  mutate(parents = unlist(map(parents, ~str_c(.x, collapse = ",")))) %>% 
  write_tsv(., file.path(sc_tbls_dir, "goterms_gp4_clusters.tsv"))

go_res[["result"]] %>% 
  mutate(parents = unlist(map(parents, ~str_c(.x, collapse = ","))),
         source = str_replace_all(source, "[[:punct:]]", " ")) %>% 
  select(-significant) %>% 
  split(., .$query) %>% 
  openxlsx::write.xlsx(., file.path(sc_tbls_dir, "goterms_gp4_cluster.xlsx"))
```
 


```{r}
plot_harmony(so, 
             "tumor_subpopulations_gp4_refined")

write_avg_expr(so,  
               "tumor_subpopulations_gp4_refined",
               file.path(sc_tbls_dir, 
                         "tumor_subpopulations_refined_avgexpr.tsv.gz"))


cell_count_mat <- get_cell_count_matrix(so, "UPN", "tumor_subpopulations_gp4_refined")

cell_count_mat[is.na(cell_count_mat)] <- 0L

cell_count_mat %>% 
  tibble::rownames_to_column("UPN") %>% 
write_csv(file.path(sc_tbls_dir, 
                         "tumor_subpopulations_refined_harmony_cell_counts.csv"))

full_marker_matrix <- full_mkrs %>% 
  top_marker_matrix(.) 
  
topn_marker_matrix <- full_mkrs %>% 
  top_marker_matrix(n = 200) 

out <- list(avg_expression = read_tsv(file.path(sc_tbls_dir, "tumor_subpopulations_refined_avgexpr.tsv.gz")),
     logFC_all = full_marker_matrix,
     top_200_logFC = topn_marker_matrix)

openxlsx::write.xlsx(out, 
                     file.path(sc_tbls_dir, "tumor_subpopulations_refined_gp4_harmony_expression_logfc_summaries.xlsx"))
```

## save

```{r}
qsave(so, file.path(obj_dir, "gp4_only.qs"), nthreads = 4)
qsave(so_neo, file.path("objects", "preprocess", "all_neoplastic.qs"))
```
  

```{r, eval = FALSE, echo = FALSE}
# for quick reloading as necessary
so <- qread(file.path(obj_dir, "gp4_only.qs"), nthreads = 4, use_alt_rep = FALSE)
so <- qread(file.path(obj_dir, "gp3_only.qs"), nthreads = 4, use_alt_rep = FALSE)
so <- qread(file.path(obj_dir, "gp34.qs"), nthreads = 4, use_alt_rep = FALSE)
```
