==> results/08_pyscenic.Rmd <==
---
title: "PyScenic"
author: "Kent Riemondy RBI"
date: "`r Sys.Date()`"
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "html") })
output:
  html_document:
    toc: true
    toc_float: true
    toc_collapsed: false
    theme: cosmo
    highlight: tango
    fig_caption: true
    code_folding: hide
    df_print: paged
---
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r libs, message=FALSE, warning=FALSE, echo=FALSE}
source(here::here("R/utils.R"))
library(presto)
```

Ran on bodhi (see [python script](../data/pyscenic/pyscenic_analysis.py))

```{r}
data_dir <- "../data/pyscenic/"

sample_dirs <- c(
  "gp34",
  "gp3_only",
  "gp4_only",
  "immune",
  "shh"
)
tf_dat <- map(sample_dirs, 
              ~read_csv(file.path(
                data_dir,
                str_c("output_", .x),
                "auc_res.csv"
              )))

names(tf_dat) <- sample_dirs
```

## SHH dataset

```{r}
so <- qread(file.path(obj_dir, "shh", "shh.qs"), use_alt_rep = FALSE)

shh_data <- tf_dat$shh %>%
  as.data.frame() %>%
  column_to_rownames("Cell") %>%
  as.matrix() %>% 
  t()

shh_data <- shh_data[, Cells(so)]
so[["TF"]] <- CreateAssayObject(data = shh_data)

Idents(so) <- "coarse_clusters_shh_harmony"
```


```{r, eval = FALSE}
DefaultAssay(so) <- "RNA"
qsave(so, file.path(obj_dir, "shh", "shh.qs"), nthreads = 4)
```


## Group 3 and Group 4 dataset

```{r}
so <- qread(file.path(obj_dir,"gp34", "gp34.qs"), nthreads = 4, use_alt_rep = FALSE)

grp34_data <- tf_dat$gp34 %>%
  as.data.frame() %>%
  column_to_rownames("Cell") %>%
  as.matrix() %>% 
  t()

grp34_data <- grp34_data[, Cells(so)]
so[["TF"]] <- CreateAssayObject(data = grp34_data)

Idents(so) <- "coarse_clusters_gp34_harmony"
DefaultAssay(so) <- "TF"

so_tf_markers <- wilcoxauc(so,
                           group_by = "coarse_clusters_gp34_harmony", 
                           seurat_assay = "TF") %>% 
  filter(logFC > 0, 
         padj < 0.05) %>% 
  group_by(group) %>% 
  arrange(padj, desc(logFC))

write_tsv(so_tf_markers, file.path(tbls_dir, "gp34", "gp34_tf_activities.tsv"))

write_markers_xlsx(split(so_tf_markers, so_tf_markers$group),
                   path = file.path(tbls_dir, "gp34", "gp34_tf_activities.xlsx"),
                   description_string = "Transcription factor regulons derived from PyScenic that are significantly enriched in each cluster")

topx <- so_tf_markers %>% 
  group_by(group) %>% 
  arrange(padj, desc(logFC), .by_group = TRUE) %>% 
  slice(1:5)

so <- ScaleData(so)

p <- DoHeatmap(so, 
               group.colors = discrete_palette_default,
        features = unique(topx$feature),
        group.by = "coarse_clusters_gp34_harmony",
        angle = 0,
        raster = FALSE, 
        draw.lines = TRUE)

p <- p +
  scale_fill_gradientn(colours = viridis::viridis(256),
                       name = "Regulon AUC\nZ-scores")

save_plot(file.path(fig_dir, "gp34", "heatmap_tfs_gp34.pdf"), 
          p,
          base_asp = 2,
          base_height = 8)

```

```{r, eval = FALSE}
DefaultAssay(so) <- "RNA"

qsave(so, file.path(obj_dir, "gp34", "gp34.qs"), nthreads = 4)
```



## Group 3 only

```{r}
dir.create(file.path(fig_dir, "gp3"), showWarnings = FALSE)
dir.create(file.path(tbls_dir, "gp3"), showWarnings = FALSE)
```

```{r}
so <- qread(file.path(obj_dir,"gp34", "gp3_only.qs"), nthreads = 4, use_alt_rep = FALSE)

# need to rerun with gp3 
grp34_data <- tf_dat$gp34 %>%
  as.data.frame() %>%
  column_to_rownames("Cell") %>%
  as.matrix() %>% 
  t()

grp34_data <- grp34_data[, Cells(so)]
so[["TF"]] <- CreateAssayObject(data = grp34_data)

```

```{r, eval = FALSE}
DefaultAssay(so) <- "RNA"
qsave(so, file.path(obj_dir, "gp34", "gp3_only.qs"), nthreads = 4)
```

## Group 4 only

```{r}
dir.create(file.path(fig_dir, "gp4"), showWarnings = FALSE)
dir.create(file.path(tbls_dir, "gp4"), showWarnings = FALSE)
```

```{r}
so <- qread(file.path(obj_dir,"gp34", "gp4_only.qs"), nthreads = 4, use_alt_rep = FALSE)

grp34_data <- tf_dat$gp4_only %>%
  as.data.frame() %>%
  column_to_rownames("Cell") %>%
  as.matrix() %>% 
  t()

grp34_data <- grp34_data[, Cells(so)]
so[["TF"]] <- CreateAssayObject(data = grp34_data)
```

```{r, eval = FALSE}
DefaultAssay(so) <- "RNA"
qsave(so, file.path(obj_dir, "gp34", "gp4_only.qs"), nthreads = 4)
```

## Immune

```{r}
so <- qread(file.path(obj_dir,"immune", "immune_so.qs"), nthreads = 4, use_alt_rep = FALSE)

grp34_data <- tf_dat$immune %>%
  as.data.frame() %>%
  column_to_rownames("Cell") %>%
  as.matrix() %>% 
  t()

grp34_data <- grp34_data[, Cells(so)]
so[["TF"]] <- CreateAssayObject(data = grp34_data)

Idents(so) <- "new_cell_type_ids"
DefaultAssay(so) <- "TF"

so_tf_markers <- wilcoxauc(so,
                           group_by = "new_cell_type_ids", 
                           seurat_assay = "TF") %>% 
  filter(logFC > 0, 
         padj < 0.05) %>% 
  group_by(group) %>% 
  arrange(padj, desc(logFC))

write_tsv(so_tf_markers, file.path(tbls_dir, "immune", "immune_tf_activities.tsv"))

write_markers_xlsx(split(so_tf_markers, so_tf_markers$group),
                   path = file.path(tbls_dir, "immune", "immune_tf_activities.xlsx"),
                   description_string = "Transcription factor regulons derived from PyScenic that are significantly enriched in each cluster")

topx <- so_tf_markers %>% 
  group_by(group) %>% 
  arrange(padj, desc(logFC), .by_group = TRUE) %>% 
  slice(1:5)

so <- ScaleData(so)

p <- DoHeatmap(so, 
               group.colors = discrete_palette_default,
        features = unique(topx$feature),
        group.by = "new_cell_type_ids",
        angle = 0,
        raster = FALSE, 
        draw.lines = TRUE)

p <- p +
  scale_fill_gradientn(colours = viridis::viridis(256),
                       name = "Regulon AUC\nZ-scores")

save_plot(file.path(fig_dir, "immune", "heatmap_tfs_immune.pdf"), 
          p,
          base_asp = 2,
          base_height = 8)

```

```{r, eval = FALSE}
DefaultAssay(so) <- "RNA"
qsave(so, file.path(obj_dir, "immune", "immune_so.qs"), nthreads = 4)
```
