==> results/09_overview.Rmd <==
---
title: "Overview"
author: "Kent Riemondy RBI"
date: "`r Sys.Date()`"
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "html") })
output:
  html_document:
    toc: true
    toc_float: true
    toc_collapsed: false
    theme: cosmo
    highlight: tango
    fig_caption: true
    code_folding: hide
    df_print: paged
---

```{r libs, message=FALSE, warning=FALSE, echo=FALSE}
source(here::here("R/utils.R"))
# rmd_dir <- "preprocess"
# 
# overview_fig_dir <- file.path(fig_dir, "overview")
# mkrs_dir <- file.path(mkrs_dir, rmd_dir)
# tbls_dir <- file.path(tbls_dir, rmd_dir)
# obj_dir <- file.path(obj_dir, rmd_dir)
# dir.create(fig_dir, showWarnings = FALSE)
library(ComplexHeatmap)
library(dendextend)
```

# QC overview

```{r}
qc_data_fig_dir <- file.path(fig_dir, "qc")
dir.create(qc_data_fig_dir, showWarnings = FALSE)

so <- qread(file.path(obj_dir, "preprocess", "unfiltered.qs"), nthreads = 4)

to_drop <- c("1008", "1192", "1202")
to_keep <- colnames(so)[which(!so@meta.data$orig.ident %in% to_drop)]
so <- subset(so, cells = to_keep)

so$subgroup_v2 <- ifelse(so$subgroup == "GP3/4", "GP3", so$subgroup)

```

### Percent Mitochondrial UMIs 

```{r}
p1 <- plot_violin(so@meta.data, 
            "UPN",
            "percent.mt",
            .fill = "subgroup_v2") +
  labs(x = "", y = "Percent UMIs from Mitochondria") + 
 geom_hline(aes(yintercept = 30), linetype = "dashed")

save_plot(file.path(qc_data_fig_dir, "prop_mito.pdf"),
          p1, base_asp = 1.75)
p1
```


### # of genes detected 

```{r}
p3 <- plot_violin(so@meta.data,
            "UPN", 
            "nFeature_RNA",
            .fill = "subgroup_v2") +
  labs(x = "", y = "# of genes per cell")

save_plot(file.path(qc_data_fig_dir, "ngene.pdf"),
          p3, base_asp = 1.75)
p3
```


### # of UMIs detected

```{r}
p2 <- plot_violin(so@meta.data, 
            "UPN",
            "nCount_RNA", 
            .fill = "subgroup_v2") +
  labs(x = "", y = "# of UMIs") + 
 geom_hline(aes(yintercept = 40000), linetype = "dashed")

save_plot(file.path(qc_data_fig_dir, "numi.pdf"),
          p2, base_asp = 1.75)
p2
```


# All data overview  

  
```{r}
all_data_fig_dir <- file.path(fig_dir, "all_data")
dir.create(all_data_fig_dir, showWarnings = FALSE)

so <- qread(file.path(obj_dir, "preprocess", "so.qs"), nthreads = 4)
so$subgroup_v2 <- ifelse(so$subgroup == "GP3/4", "GP3", so$subgroup)
```


## Make overview umaps

### Subgroup

```{r}
n_cells <- ncol(so) %>% scales::comma() %>% str_c(., " cells")
p <- plot_umap(so, "subgroup_v2", 
          .col = palette_OkabeIto, 
          label_text = TRUE,
          legend_title = "",
          label_col = "black",
          sorted = "random", 
          ggrepel_opts = list(force = 0)) +
  annotate(x = 17,
           y = -18, 
           geom = "text",
           label = n_cells)
  
save_plot(file.path(all_data_fig_dir, "umap_by_subgroup.pdf"),
          p, 
          base_asp = 1.4)

p
```

### UPN

```{r}
p <- plot_umap(so, "UPN", sorted = "random", legend_title = "") + 
  theme( legend.text = element_text(size = 8))

save_plot(file.path(all_data_fig_dir, "umap_by_UPN.pdf"),
          p, 
          base_asp = 1.4)

p
```

### UPN + cell type + subgroup

Color by UPN, annotate with text cell type or subgroup

```{r}
plt_text <- get_metadata(so) %>% 
  select(coarse_cell_type, subgroup_v2, starts_with('UMAP')) %>% 
  mutate(txt = ifelse(coarse_cell_type == "malignant",
                      subgroup_v2,
                      coarse_cell_type),
         txt = case_when(
           txt == "lymphocytes" ~ "lymph",
           txt == "macrophage_monocytes" ~ "myel",
           txt == "oligodendrocytes_astrocytes_other" ~ "oligo/astro",
           TRUE ~ txt 
         )) %>% 
  group_by(txt) %>% 
  mutate(txt_label = ifelse(
    scbp:::closest_to_point(data.frame(.data[["UMAP_1"]],
                                .data[["UMAP_2"]]),
                               c(median(UMAP_1),
                                 median(UMAP_2))),
    txt,
    ""
  )) 

p <- plot_umap(so, "UPN", sorted = "random", legend_title = "") + 
  ggrepel::geom_text_repel(data = plt_text,
                           min.segment.length = Inf, 
                           force = 0.2, 
                           point.padding = unit(0.5, "line"),
                           aes(label = txt_label),
                           seed = 1) + 
  guides(color = guide_legend(ncol = 1, override.aes = list(size = 1))) +
  annotate(x = -15,
           y = -18, 
           size = 3,
           geom = "text",
           label = n_cells) +
  theme(legend.text = element_text(size = 8),
        legend.key.size = unit(0.5, 'line'))


save_plot(file.path(all_data_fig_dir, "umap_by_UPN_label_subgroup.pdf"),
          p,
          base_asp = 1.4)
p
# for some reason the plot printed by rstudio is different than the saved plot
# unable to fix with messing around with seeds, save_plot, ggsave, or pdf...
```

```{r}
p <- plot_umap(so, "UPN", sorted = "random", legend_title = "") + 
  ggrepel::geom_text_repel(data = plt_text,
                           min.segment.length = Inf, 
                           force = 0.2, 
                           point.padding = unit(0.5, "line"),
                           aes(label = txt_label),
                           seed = 1) + 
  guides(color = guide_legend(ncol = 1, override.aes = list(size = 1))) +
  annotate(x = -15,
           y = -18, 
           size = 3,
           geom = "text",
           label = n_cells) +
  theme(legend.text = element_text(size = 8),
        legend.key.size = unit(0.5, 'line'),
        axis.text = element_blank(),
        axis.title = element_blank(),
        axis.line = element_blank(),
        axis.ticks = element_blank())

p

save_plot(file.path(all_data_fig_dir, "umap_by_UPN_label_subgroup_no_axes.pdf"),
          p,
          base_asp = 1.4)
```


### coarse clusters

```{r}
p <- plot_umap(so, "coarse_clusters") + 
  theme( legend.text = element_text(size = 6))

save_plot(file.path(all_data_fig_dir, "umap_by_cluster.pdf"),
          p, 
          base_asp = 1.4)
```

### coarse cell type

```{r}
coarse_cell_type_pretty_id <- c("Neoplastic",
                               "Lymphocytes",
                               "Glia",
                               "Myeloid")

names(coarse_cell_type_pretty_id) <- c("malignant",
                               "lymphocytes",
                               "oligodendrocytes_astrocytes_other",
                               "macrophage_monocytes")

so$coarse_cell_type_id <- coarse_cell_type_pretty_id[so$coarse_cell_type]
p <- plot_umap(so, "coarse_cell_type_id", .cols = palette_OkabeIto, legend_title = "") 

p 

save_plot(file.path(all_data_fig_dir, "umap_by_coarse_cell_type.pdf"),
          p, 
          base_asp = 1.4)
```

## Plot cell distibutions per subgroup and tumor

```{r}
p4 <- plot_cell_proportions(so, "UPN", "coarse_cell_type_id", facet_by = "subgroup_v2") +
   labs(x = "Tumor ID",
       y = "Proportion of each cell type") +
  scale_fill_manual(values = rev(palette_OkabeIto[1:4]), 
                    name = "") +
   theme( strip.text = element_text(angle = 90))


save_plot(file.path(all_data_fig_dir, "cell_types_per_tumor.pdf"), 
          p4,
          base_width = 7,
          base_aspect_ratio = 1.5)
p4

```

Combine with QC plots from above
```{r}
p <- plot_grid(p1, p2, p3, p4, nrow = 2, ncol = 2, scale = 0.975)

save_plot(file.path(fig_dir, "qc", "combined_metrics.pdf"), 
          p, nrow = 2, ncol = 2, base_asp = 1.5)
p
```

## infer CNV figures

run on server per subgroup. Consider replotting heatmaps with better visualizations. 

See [11_cnv.Rmd](11_cnv.Rmd)

## Cell counts

```{r}
get_metadata(so) %>% 
  count(UPN, subgroup_v2, coarse_cell_type) %>% 
  rename(subgroup = subgroup_v2) %>% 
  pivot_wider(names_from = "coarse_cell_type",
              values_from = "n", 
              values_fill = 0L) %>% 
  write_csv(., file.path(tbls_dir, "preprocess", "cell_counts_coarse_cell_type.csv"))
```

# All neoplastic 
## Subtypes based on Cavalli paper

```{r}
so <- qread(file.path(obj_dir, "preprocess", "all_neoplastic.qs"))

so$subgroup_v2 <- ifelse(so$subgroup == "GP3/4", "GP3", so$subgroup)
# drop excluded subpopulations

so$tmp_subpopulation <- ifelse(is.na(so$tumor_subpopulation), "WNT", so$tumor_subpopulation)
to_keep <- colnames(so)[!str_detect(so$tmp_subpopulation, "X")]

so <- subset(so, cells = to_keep)

all_neo_fig_dir <- file.path(fig_dir, "all_neoplasic")
dir.create(all_neo_fig_dir, showWarnings = FALSE)
```

```{r}
n_cells <- ncol(so) %>% scales::comma() %>% str_c(., " cells")
p <- plot_umap(so, "subgroup_v2", .cols = palette_OkabeIto,
               legend_title = "") + 
   annotate(x = -12,
           y = -20,
           size = 3,
           geom = "text",
           label = n_cells)
p
save_plot(file.path(all_neo_fig_dir, "umap_neoplastic_by_subgroup.pdf"),
          p, 
          base_asp = 1.2)

p234 <- plot_umap(so, 
               c("NRL", "IMPG2", "tmp_subpopulation"))

p234[[3]] <- p234[[3]] +
    guides(color = guide_legend(ncol = 1, override.aes = list(size = 1))) +
  theme(legend.text = element_text(size = 8),
        legend.title = element_blank(),
        legend.key.size = unit(0.5, 'line'))

p <-  p +
    guides(color = guide_legend(ncol = 1, override.aes = list(size = 1))) +
  theme(legend.text = element_text(size = 8),
        legend.title = element_blank(),
        legend.key.size = unit(0.5, 'line'))

p <- append(list(p), p234)
p <- plot_grid(plotlist = p, nrow = 1, ncol = 4)
save_plot(file.path(all_neo_fig_dir, "umap_subpopulations_and_photoreceptor_markers.pdf"),
          p, 
          nrow = 1,
          ncol = 4,
          base_asp = 1.2)

p <- plot_umap(so, "methyl_nCNV",
               legend_title = "") 
p
save_plot(file.path(all_neo_fig_dir, "umap_neoplastic_by_cnvs.pdf"),
          p, 
          base_asp = 1.2)

subtype_pal <- c(
  brewer.pal(4, "Blues")[2:4],
  brewer.pal(4, "Greens")[2:4],
  brewer.pal(4, "Reds")[2:4],
  brewer.pal(4, "Purples")[c(2,4)]
)
p <- plot_umap(so, "subtype", .cols = subtype_pal,
               legend_title = "") 
p
save_plot(file.path(all_neo_fig_dir, "umap_neoplastic_by_subtype.pdf"),
          p, 
          base_asp = 1.5)


p <- plot_umap(so, "UPN", legend_title = "") +
  theme( legend.text = element_text(size = 8))

save_plot(file.path(all_neo_fig_dir, "umap_neoplastic_by_UPN.pdf"),
          p, 
          base_asp = 1.4)
```


### Cell cycle per subpopulation

```{r}
p <- get_metadata(so) %>% 
  select(cell, tumor_subpopulation, Phase) %>% 
  group_by(tumor_subpopulation) %>% 
  mutate(cells_per_subpopulation = n()) %>% 
  group_by(tumor_subpopulation, Phase) %>%
  mutate(cells_per_phase = n(),
            perc = 100 * (cells_per_phase / cells_per_subpopulation)) %>% 
  select(-cell) %>% 
  na.omit() %>% 
  unique() %>% 
  arrange(tumor_subpopulation) %>% 
  mutate(Phase = factor(Phase, levels = c("G1", "S", "G2M"))) %>% 
  mutate(subgroup = str_split(tumor_subpopulation, "-", simplify = T) %>% .[, 1],
         subpopulation = str_split(tumor_subpopulation, "-", simplify = T) %>%
           .[, 2] %>% 
           str_sub(1, 1)) %>% 
  ggplot(aes(tumor_subpopulation, perc)) +
    geom_col(aes(fill = Phase)) +
    scale_fill_manual(values = palette_OkabeIto) + 
    facet_grid(~subpopulation, scales = "free_x") +
    labs(x = "Subpopulation",
         y = "Percent of cells") +
  theme(axis.text.x = element_text(angle =90,
                                   hjust = 1,
                                   vjust = 0.5),
        strip.background = element_blank(),
        strip.text = element_blank())

p 

save_plot(file.path(all_neo_fig_dir, "percent_cycling_per_subgroup.pdf"),
          p,
          base_asp = 1.6)
```


# SHH tumors

```{r}
so <- qread(file.path("objects", "shh", "shh.qs"))

shh_fig_dir <- file.path(fig_dir, "shh")
dir.create(shh_fig_dir, showWarnings = FALSE)
```

### Various genes in SHH


```{r}
famous_genes <- c(
  "MYCN",
  "GLI2",
  "SOX2",
  "BOC"
)

plts <- map(famous_genes,
    ~plot_harmony(so, .x, minimal_theme = TRUE))

names(plts) <- famous_genes

plot_grid(plotlist = plts)

iwalk(plts,
     ~save_plot(file.path(shh_fig_dir, 
                          paste0("umap_neoplastic_by_", .y, ".pdf")),
          .x, 
          base_asp = 1.2))
```


### Harmonized projection

Using alignment techniques generates marker gene lists that are prettty similar to those seen by NMF/scHPF. Also produces easier to digest visualizations via UMAP rather than heatmaps.

```{r}
p <- plot_harmony(so, 
             "coarse_clusters_shh_harmony", 
             label_color = "black", 
             label_text = TRUE,
             legend_title = "")

save_plot(file.path(shh_fig_dir, "umap_shh_by_clusters.pdf"),
          p, 
          base_asp = 1.4)

p <- plot_harmony(so, 
             "UPN", 
             legend_title = "")

save_plot(file.path(shh_fig_dir, "umap_shh_by_upn.pdf"),
          p, 
          base_asp = 1.4)

p <- plot_harmony(so, 
             "Phase", 
             .cols = palette_OkabeIto,
             legend_title = "")

save_plot(file.path(shh_fig_dir, "umap_shh_by_cc_phase.pdf"),
          p, 
          base_asp = 1.4)

p <- plot_harmony(so, 
             "subtype", 
             .cols = palette_OkabeIto,
             legend_title = "") +
  scale_color_manual(values = palette_OkabeIto, 
                     drop = TRUE,
                     name = "")

save_plot(file.path(shh_fig_dir, "umap_shh_by_subtype.pdf"),
          p, 
          base_asp = 1.5)

```

### overview plot

```{r}
# drop contamination populations
to_keep <- colnames(so)[!str_detect(so$tumor_subpopulations_shh, "X")]
so_sub <- subset(so, cells = to_keep)
```

```{r}
text_to_show <- tribble(
  ~tumor_subpopulations_shh, ~sub_pop_long, ~go, ~tf, ~genes,
  "SHH-A1", "SHH-A mitotic G2M", "GO: G2M and S phase", "TF: MYBL2", "",
  "SHH-A2", "SHH-A mitotic S", "GO: G2M and S phase", "TF: E2Fs", "",
  "SHH-B1", "SHH-B progenitor 1", "GO: Regulation of neuron\ndifferentiation/Translation", "TF: ATOH1,ZEB1,HES6", "Genes: ATOH1,SFRP1,\nSOX2,SOX9,Ribosomal",
  "SHH-B2", "SHH-B progenitor 2", "GO: Translation", "TF: ELK1,THAP11,MAX", "Genes: Ribosomal",
  "SHH-C1", "SHH-C differentiated 1", "GO: RNA processing/axon-dendritic\ntransport", "TF: POU2F2,KMT2A", "Genes: GRIA2,LUC7L3,DST",
  "SHH-C2", "SHH-C differentiated 2", "GO: Axon development", "TF: SOX4,NEUROD1", "Genes: STMN2/4",
)

find_max_with_limits <- function(x, q = c(0.1, 0.9)){
  lims <- quantile(x, q)
  vals <- x[x > lims[1] & x < lims[2]]
  vals[abs(vals) == max(abs(vals))]
}


plt_dat <- get_metadata(so_sub) %>%
  select(`UMAP_1.1`, `UMAP_2.1`, tumor_subpopulations_shh) %>%
    left_join(text_to_show, by = "tumor_subpopulations_shh") %>%
  mutate(txt_label = str_c(sub_pop_long,
                     go,
                     tf,
                     genes,
                     sep = "\n")) %>%
  group_by(tumor_subpopulations_shh, txt_label) %>%
  summarize(UMAPHarmony_1 = find_max_with_limits(`UMAP_1.1`),
            UMAPHarmony_2 = find_max_with_limits(`UMAP_2.1`))

lval <- 8
p <- plot_harmony(so_sub, "tumor_subpopulations_shh") +
  theme(legend.position = "none") + 
 
  ggrepel::geom_label_repel(data = plt_dat,
                            aes(x = UMAPHarmony_1,
                                y = UMAPHarmony_2,
                                label = txt_label),
                            size = 3, 
                            box.padding =0.1,
                           force = 0.05, 
                           max.overlaps = 2,
                           max.time = 2.0,
                           xlim = c(-lval - 1.5, lval - 1.5),
                           ylim = c(-lval, lval),
                           hjust = 0,
                        fill = alpha("white", 0.5),
                        seed = 42,
                           label.size = 0.5, 
                        direction = "both") +
  coord_cartesian(clip = "off") 

save_plot(file.path(shh_fig_dir, "umap_shh_harmony_overview.pdf"),
          p, 
          base_height = 6,
          base_asp = 1.2)
```

```{r}
p <- plot_harmony(so_sub,
                  "tumor_subpopulations_shh",
                  pt_size = 0.2,
                  label_text = TRUE,
                  label_color = "black",
                  ggrepel_opts = list(size = 8, force = 0.01)) +
  theme(legend.position = "none")

p

save_plot(file.path(shh_fig_dir, "umap_shh_harmony_overview_simple_v2.pdf"),
          p, 
          base_height = 6,
          base_asp = 1.2)
```
```{r}
 plot_harmony(so_sub,
                  "tumor_subpopulations_shh")
```

### harmony cluster markers

```{r fig.height = 6}
mkrs <- read_tsv(file.path(mkrs_dir, "shh", "harmony_markers_shh.tsv"))

topx <- mkrs %>% 
  arrange(padj) %>% 
  group_by(group) %>% 
  slice(1:10) 

p <- DotPlot(so, 
        features = unique(topx$feature),
        cols = "Spectral",
        group.by = "coarse_clusters_shh_harmony") +
  coord_flip() +
  labs(y = "Cluster",
       x = "") +
  theme(axis.text.x = element_text(size = 24),
        axis.text.y = element_text(size = 12),
        axis.title.x = element_text(size = 24))

p

save_plot(file.path(shh_fig_dir, 
                    "dotplot_shh_markers.pdf"), 
          p, base_height = 12, base_asp = 1)


```

### Northcutt metaprograms

Predict that Harmony + cluster ~= NMF per sample + clustering.

Yep. 

```{r}
shh_mps <- c(
  "SHH-A-cell_cycle",
  "SHH-B-SHH_signaling",
  "SHH-C-Differentiation"
)

names(shh_mps) <- paste0("Hovestadt_et_al_", shh_mps)

plts <- imap(shh_mps,
            ~plot_harmony(so, 
               .x, 
               show_negative = TRUE,
               minimal_theme = TRUE
               ) + 
              labs(title = .y))

out_ids <- names(plts) %>% 
  str_split(., "-") %>% 
  map_chr(., ~str_c(.x[1:2], collapse = "-"))

walk2(plts, out_ids, 
    ~save_plot(file.path(shh_fig_dir, 
                         paste0("umap_shh_by_", .y, ".pdf")),
          .x, 
          base_asp = 1.4))

```

### plot putative stem markers from mouse SHH Cancel cell paper

Most of the tSNE plots show really sparse and low expression of these genes. Not sure how reliable those finding are. 

```{r}
stem_mkrs <- c(
  "PROM1",
  "SOX2",
  "OLIG1",
  "OLIG2",
  "NES",
  "HES1",
  "MYCN",
  "BOC",
  "GLI2"
)

plts <- plot_harmony(so, 
               stem_mkrs, 
               minimal_theme = TRUE) 
names(plts) <- stem_mkrs
iwalk(plts, 
    ~save_plot(file.path(shh_fig_dir, 
                         paste0("umap_shh_by_", .y, ".pdf")),
          .x, 
          base_asp = 1.4))
```


### plot some CNV genes



```{r}
cnv_mkrs <- c(
  "MYCN",
  "TERT",
  "CDKN2A"
)

plts <- map(cnv_mkrs,
            ~plot_feature(so, 
               .x, 
               embedding = "umap", 
               minimal_theme = TRUE))

p <- plot_umap(so, 
               "UPN", 
               minimal_theme = TRUE,
               label_text = TRUE)


plts <- c(list(p), plts)
names(plts) <- c("UPN", cnv_mkrs)
iwalk(plts, 
    ~save_plot(file.path(shh_fig_dir, 
                         paste0("umap_shh_by_", .y, ".pdf")),
          .x, 
          base_asp = 1))
```

### cell count tables 

```{r}
res <- so@meta.data %>% 
  group_by(coarse_clusters_shh_harmony, UPN) %>% 
  summarize(n = n()) %>% 
  ungroup() %>% 
  spread(UPN, n, fill = 0L) 

write_csv(res, file.path(tbls_dir, "shh", "shh_cells_per_cluster_per_sample.csv"))

res <- so@meta.data %>% 
    group_by(coarse_clusters_shh_harmony, UPN) %>% 
    summarize(n = n()) %>%
  group_by(coarse_clusters_shh_harmony) %>% 
  mutate(total = sum(n),
         percent = 100 * n / total) %>% 
  ungroup() %>% 
  select(-n, -total) %>% 
  spread(UPN, percent, fill = 0L) 

  
write_csv(res, file.path(tbls_dir,"shh",  "shh_percent_cells_per_cluster_per_sample.csv"))

```




# GP3  tumors

```{r}
so <- qread(file.path("objects", "gp34", "gp3_only.qs"))

gp3_fig_dir <- file.path(fig_dir, "gp3")
dir.create(gp3_fig_dir, showWarnings = FALSE)
```

### Various marker genes for Gp3 and Gp4

```{r}
famous_genes <- c(
  "EOMES",
  "LEMD1",
  "UNC5D",
  "EPHA8"
)

plts <- map(famous_genes,
    ~plot_umap(so, .x, minimal_theme = TRUE))

names(plts) <- famous_genes

plot_grid(plotlist = plts)

iwalk(plts,
     ~save_plot(file.path(gp3_fig_dir, 
                          paste0("umap_gp3_by_", .y, ".pdf")),
          .x, 
          base_asp = 1.2))
```


### Harmonized projection

Using alignment techniques generates marker gene lists that are prettty similar to those seen by NMF/scHPF. Also produces easier to digest visualizations via UMAP rather than heatmaps.

```{r}
p <- plot_harmony(so, 
             "coarse_clusters_gp34_harmony", 
             label_color = "black", 
             label_text = TRUE,
             legend_title = "",
             sorted = "random") +
  labs(x = "UMAP 1",
       y = "UMAP 2")

save_plot(file.path(gp3_fig_dir, "umap_gp3_by_clusters.pdf"),
          p, 
          base_asp = 1.4)

p <- plot_harmony(so, 
             "UPN", 
             legend_title = "",
             sorted = "random") +
  labs(x = "UMAP 1",
       y = "UMAP 2")

save_plot(file.path(gp3_fig_dir, "umap_gp3_by_upn.pdf"),
          p, 
          base_asp = 1.4)

p <- plot_harmony(so, 
             "Phase", 
             .cols = palette_OkabeIto,
              legend_title = "",
             sorted = "random") +
  labs(x = "UMAP 1",
       y = "UMAP 2")

save_plot(file.path(gp3_fig_dir, "umap_gp3_by_cc_phase.pdf"),
          p, 
          base_asp = 1.4)

p <- plot_harmony(so, 
             "subtype", 
             .cols = palette_OkabeIto,
              legend_title = "",
             sorted = "random") +
  scale_color_manual(values = palette_OkabeIto, 
                     drop = TRUE) +
  labs(x = "UMAP 1",
       y = "UMAP 2")

save_plot(file.path(gp3_fig_dir, "umap_gp3_by_subtype.pdf"),
          p, 
          base_asp = 1.5)


p <- plot_harmony(so, 
             "tumor_cell_type_gp3", 
             .cols = palette_OkabeIto,
              legend_title = "",
             sorted = "random") +
  labs(x = "UMAP 1",
       y = "UMAP 2")

save_plot(file.path(gp3_fig_dir, "umap_gp3_by_tumor_cell_type.pdf"),
          p, 
          base_asp = 1.5)
```


### overview plot

```{r}
text_to_show <- tribble(
  ~tumor_subpopulations_gp3, ~sub_pop_long, ~go, ~tf, ~genes,
  "GP3-A", "Gp3-A mitotic", "GO: G2M and S phase", "TF: MYBL2", "",
  "GP3-B1", "Gp3-B progenitor 1", "GO: Translation/Oxidative Phosphorylation", "TF: HMGA1, MYC, MAX", "Genes: Ribosomal",
  "GP3-B2", "Gp3-B progenitor 2", "GO: Translation/Oxidative Phosphorylation", "TF: HLX, ATF4, MYC", "Genes: Ribosomal, MYC",
  "GP3-C1", "Gp3-C differentiated 1", "GO: RNA processing/axon-dendritic\ntransport", "TF: ARNT", "Genes: LUC7L3,DST",
  "GP3-C2", "Gp3-C differentiated 2", "GO: Photoreceptor", "TF: SHOX2,CRX", "Genes: NRL, IMPG2",
)

find_max_with_limits <- function(x, q = c(0.1, 0.9)){
  lims <- quantile(x, q)
  vals <- x[x > lims[1] & x < lims[2]]
  vals[abs(vals) == max(abs(vals))]
}

# plt_dat <- get_metadata(so) %>% 
#   select(UMAPHarmony_1, UMAPHarmony_2, tumor_subpopulations_gp3) %>% 
#     left_join(text_to_show, by = "tumor_subpopulations_gp3") %>% 
#   mutate(txt = str_c(sub_pop_long,
#                      go,
#                      tf,
#                      genes,
#                      sep = "\n")) %>% 
#   group_by(tumor_subpopulations_gp3) %>%
#   sample_n(100) %>% 
#   mutate(txt_label = ifelse(
#     scbp:::closest_to_point(data.frame(.data[["UMAPHarmony_1"]],
#                                 .data[["UMAPHarmony_2"]]),
#                                c(find_max_with_limits(UMAPHarmony_1),
#                                  find_max_with_limits(UMAPHarmony_2))),
#     txt,
#     ""
#   )) %>% 
#   ungroup()
# 
plt_dat <- get_metadata(so) %>%
  select(UMAPHarmony_1, UMAPHarmony_2, tumor_subpopulations_gp3) %>%
    left_join(text_to_show, by = "tumor_subpopulations_gp3") %>%
  mutate(txt_label = str_c(sub_pop_long,
                     go,
                     tf,
                     genes,
                     sep = "\n")) %>%
  group_by(tumor_subpopulations_gp3, txt_label) %>%
  summarize(UMAPHarmony_1 = find_max_with_limits(UMAPHarmony_1),
            UMAPHarmony_2 = find_max_with_limits(UMAPHarmony_2))

lval <- 8.75
p <- plot_harmony(so, "tumor_subpopulations_gp3") +
  theme(legend.position = "none") + 
 
  ggrepel::geom_label_repel(data = plt_dat,
                            aes(x = UMAPHarmony_1,
                                y = UMAPHarmony_2,
                                label = txt_label),
                            size = 3, 
                            box.padding =0.1,
                           force = 0.02, 
                           max.overlaps = 1,
                           max.time = 2.0,
                           xlim = c(-lval, lval),
                           ylim = c(-lval, lval),
                           hjust = 0,
                        fill = alpha("white", 0.5),
                        seed = 42,
                           label.size = 0.5, 
                        direction = "both") +
  coord_cartesian(clip = "off") 

save_plot(file.path(gp3_fig_dir, "umap_gp3_harmony_overview.pdf"),
          p, 
          base_height = 6,
          base_asp = 1.2)
```

```{r}
p <- plot_harmony(so,
                  "tumor_subpopulations_gp3",
                  pt_size = 0.2,
                  label_text = TRUE,
                  label_color = "black",
                  ggrepel_opts = list(size = 8, force = 0)) +
  labs(x = "UMAP 1",
       y = "UMAP 2") + 
  theme(legend.position = "none")

p

save_plot(file.path(gp3_fig_dir, "umap_gp3_harmony_overview_simple.pdf"),
          p, 
          base_height = 6,
          base_asp = 1.2)
```

### harmony cluster markers

```{r fig.height = 6}
mkrs <- read_tsv(file.path(mkrs_dir, "gp34", "harmony_markers_gp3_only.tsv"))

topx <- mkrs %>% 
  arrange(padj) %>% 
  group_by(group) %>% 
  slice(1:10) 

p <- DotPlot(so, 
        features = unique(topx$feature),
        cols = "Spectral",
        group.by = "coarse_clusters_gp3_harmony") +
  coord_flip() +
  labs(y = "Cluster",
       x = "") +
  theme(axis.text.x = element_text(size = 24),
        axis.text.y = element_text(size = 12),
        axis.title.x = element_text(size = 24))

p

save_plot(file.path(gp3_fig_dir, 
                    "dotplot_gp3_markers.pdf"), 
          p, base_height = 12, base_asp = 0.75)


```

### Northcutt metaprograms


```{r}
mp_ids <- c(
  "Group 3/4-A-cell_cycle",
  "Group 3/4-B-Translation_Myc",
  "Group 3/4-C-Differentiation"
)

names(mp_ids) <- paste0("Hovestadt_et_al_", mp_ids)

plts <- imap(mp_ids,
            ~plot_harmony(so, 
               .x, 
               show_negative = TRUE,
               minimal_theme = TRUE) + 
              labs(title = .y))

out_ids <- names(plts) %>% 
  str_split(., "-") %>% 
  map_chr(., ~str_c(.x[1:2], collapse = "-")) %>% 
  str_replace("3/4", "3 and 4")

walk2(plts, out_ids, 
    ~save_plot(file.path(gp3_fig_dir, 
                         paste0("umap_gp3_by_", .y, ".pdf")),
          .x, 
          base_asp = 1))

```

### plot various gnes

Most of the tSNE plots show really sparse and low expression of these genes. Not sure how reliable those finding are. 

```{r}

famous_genes <- c(
  "EOMES",
  "MYC",
  "NRL",
  "TOP2A",
  "CRX",
  "ROM1",
  "GRIA2"
)

plts <- plot_harmony(so, 
               famous_genes, 
               minimal_theme = TRUE)
names(plts) <- famous_genes
iwalk(plts, 
    ~save_plot(file.path(gp3_fig_dir, 
                         paste0("umap_gp3_by_", .y, ".pdf")),
          .x, 
          base_asp = 1))
```


### cell count tables 


```{r}
res <- so@meta.data %>% 
  group_by(coarse_clusters_gp3_harmony, UPN) %>% 
  summarize(n = n()) %>% 
  ungroup() %>% 
  spread(UPN, n, fill = 0L) 

write_csv(res, file.path(tbls_dir, "gp3",  "gp3_cells_per_cluster_per_sample.csv"))

res <- so@meta.data %>% 
    group_by(coarse_clusters_gp3_harmony, UPN) %>% 
    summarize(n = n()) %>%
  group_by(coarse_clusters_gp3_harmony) %>% 
  mutate(total = sum(n),
         percent = 100 * n / total) %>% 
  ungroup() %>% 
  select(-n, -total) %>% 
  spread(UPN, percent, fill = 0L) 

  
write_csv(res, file.path(tbls_dir, "gp3",
                  "gp3_percent_cells_per_cluster_per_sample.csv"))

res <- so@meta.data %>%
    group_by(coarse_clusters_gp3_harmony, subtype) %>%
  summarize(n = n()) %>%
  ungroup() %>%
  spread(subtype, n, fill = 0L)


write_csv(res, file.path(tbls_dir, "gp3",
                  "gp3_subtypes_per_cluster.csv"))

res <- so@meta.data %>%
    group_by(coarse_clusters_gp3_harmony, subgroup) %>%
  summarize(n = n()) %>%
  ungroup() %>%
  spread(subgroup, n, fill = 0L)


write_csv(res, file.path(tbls_dir, "gp3",
                  "gp3_subgroups_per_cluster.csv"))
  
```

# GP4  tumors

```{r}
so <- qread(file.path("objects", "gp34", "gp4_only.qs"))

gp4_fig_dir <- file.path(fig_dir, "gp4")
dir.create(gp4_fig_dir, showWarnings = FALSE)
```

### Various marker genes for Gp3 and Gp4

```{r}
famous_genes <- c(
  "EOMES",
  "LEMD1",
  "UNC5D",
  "EPHA8"
)

plts <- map(famous_genes,
    ~plot_umap(so, .x, minimal_theme = TRUE))

names(plts) <- famous_genes

plot_grid(plotlist = plts)

iwalk(plts,
     ~save_plot(file.path(gp4_fig_dir, 
                          paste0("umap_gp4_by_", .y, ".pdf")),
          .x, 
          base_asp = 1.2))
```


### Harmonized projection

Using alignment techniques generates marker gene lists that are prettty similar to those seen by NMF/scHPF. Also produces easier to digest visualizations via UMAP rather than heatmaps.

```{r}
p <- plot_harmony(so, 
             "coarse_clusters_gp4_harmony", 
             label_color = "black", 
             label_text = TRUE,
             legend_title = "",
             sorted = "random") +
  labs(x = "UMAP 1",
       y = "UMAP 2")

save_plot(file.path(gp4_fig_dir, "umap_gp4_by_clusters.pdf"),
          p, 
          base_asp = 1.4)

p <- plot_harmony(so, 
             "UPN", 
             legend_title = "",
             sorted = "random") +
  labs(x = "UMAP 1",
       y = "UMAP 2")

save_plot(file.path(gp4_fig_dir, "umap_gp4_by_upn.pdf"),
          p, 
          base_asp = 1.4)

p <- plot_harmony(so, 
             "Phase", 
             .cols = palette_OkabeIto,
              legend_title = "",
             sorted = "random") +
  labs(x = "UMAP 1",
       y = "UMAP 2")

save_plot(file.path(gp4_fig_dir, "umap_gp4_by_cc_phase.pdf"),
          p, 
          base_asp = 1.4)

p <- plot_harmony(so, 
             "subtype", 
             .cols = palette_OkabeIto,
              legend_title = "",
             sorted = "random") +
  scale_color_manual(values = palette_OkabeIto, 
                     drop = TRUE) +
  labs(x = "UMAP 1",
       y = "UMAP 2")

save_plot(file.path(gp4_fig_dir, "umap_gp4_by_subtype.pdf"),
          p, 
          base_asp = 1.5)


p <- plot_harmony(so, 
             "tumor_cell_type_gp4", 
              legend_title = "",
             sorted = "random") +
  labs(x = "UMAP 1",
       y = "UMAP 2")

save_plot(file.path(gp4_fig_dir, "umap_gp4_by_tumor_cell_type.pdf"),
          p, 
          base_asp = 1.5)
```


### overview plot

```{r}
# drop contamination populations
to_keep <- colnames(so)[!str_detect(so$tumor_subpopulations_gp4, "X")]
so_sub <- subset(so, cells = to_keep)
```

```{r}
text_to_show <- tribble(
  ~tumor_subpopulations_gp4, ~sub_pop_long, ~go, ~tf, ~genes,
  "GP4-A1", "Gp4-A mitotic G2M", "GO: G2M", "TF: FOXM1, MYBL2", "",
  "GP4-A2", "Gp4-A mitotic S", "GO: S phase", "TF: ZNF492, E2Fs", "",
  "GP4-B1", "Gp4-B progenitor 1", "GO: Translation/Oxidative Phosphorylation", "TF: BARHL1,NEURDOD2,SOX4,\nJUNB,JUND", "Genes: Ribosomal",
  "GP4-B2", "Gp4-B progenitor 2", "GO: Translation", "TF: HEY, TEAD2 ELK1", "Genes: STMN2, Ribosomal",
  "GP4-C1", "Gp4-C differentiated 1", "GO: RNA processing/axon-dendritic\ntransport", "TF: ONECUT2,LHX4", "Genes: GRIA2,LUC7L3,DST",
  "GP4-C2", "Gp4-C differentiated 2", "GO: Photoreceptor", "TF: MEIS1,CRX", "Genes: NRL, IMPG2",
)

plt_dat <- get_metadata(so_sub) %>%
  select(`UMAP_1.1`, `UMAP_2.1`, tumor_subpopulations_gp4) %>%
    left_join(text_to_show, by = "tumor_subpopulations_gp4") %>%
  mutate(txt_label = str_c(sub_pop_long,
                     go,
                     tf,
                     genes,
                     sep = "\n")) %>%
  group_by(tumor_subpopulations_gp4, txt_label) %>%
  summarize(UMAPHarmony_1 = find_max_with_limits(`UMAP_1.1`),
            UMAPHarmony_2 = find_max_with_limits(`UMAP_2.1`))

lval <- 10
p <- plot_harmony(so_sub, "tumor_subpopulations_gp4") +
  theme(legend.position = "none") + 

  ggrepel::geom_label_repel(data = plt_dat,
                            aes(x = UMAPHarmony_1,
                                y = UMAPHarmony_2,
                                label = txt_label),
                            size = 3, 
                            box.padding =0.1,
                           force = 0.02, 
                           max.overlaps = 2,
                           max.time = 2.0,
                           xlim = c(-lval, lval),
                           ylim = c(-lval, lval),
                           hjust = 0,
                        fill = alpha("white", 0.5),
                        seed = 42,
                           label.size = 0.5, 
                        direction = "both") +
  coord_cartesian(clip = "off") 

save_plot(file.path(gp4_fig_dir, "umap_gp4_harmony_overview.pdf"),
          p, 
          base_height = 6,
          base_asp = 1.2)
```

```{r}
p <- plot_harmony(so_sub,
                  "tumor_subpopulations_gp4",
                  pt_size = 0.2,
                  label_text = TRUE,
                  label_color = "black",
                  ggrepel_opts = list(size = 8, force = 0.13)) +
  labs(x = "UMAP 1",
       y = "UMAP 2") + 
  theme(legend.position = "none")

p

save_plot(file.path(gp4_fig_dir, "umap_gp4_harmony_overview_simple.pdf"),
          p, 
          base_height = 6,
          base_asp = 1.2)
```


### harmony cluster markers

```{r fig.height = 6}
mkrs <- read_tsv(file.path(mkrs_dir, "gp34", "harmony_markers_gp4_only.tsv"))

topx <- mkrs %>% 
  arrange(padj) %>% 
  group_by(group) %>% 
  slice(1:10) 

p <- DotPlot(so, 
        features = unique(topx$feature),
        cols = "Spectral",
        group.by = "coarse_clusters_gp4_harmony") +
  coord_flip() +
  labs(y = "Cluster",
       x = "") +
  theme(axis.text.x = element_text(size = 24),
        axis.text.y = element_text(size = 12),
        axis.title.x = element_text(size = 24))

p

save_plot(file.path(gp4_fig_dir, 
                    "dotplot_gp4_markers.pdf"), 
          p, base_height = 12, base_asp = 0.75)


```

### Northcutt metaprograms


```{r}
mp_ids <- c(
  "Group 3/4-A-cell_cycle",
  "Group 3/4-B-Translation_Myc",
  "Group 3/4-C-Differentiation"
)

names(mp_ids) <- paste0("Hovestadt_et_al_", mp_ids)

plts <- imap(mp_ids,
            ~plot_harmony(so, 
               .x, 
               show_negative = TRUE,
               minimal_theme = TRUE) + 
              labs(title = .y))

out_ids <- names(plts) %>% 
  str_split(., "-") %>% 
  map_chr(., ~str_c(.x[1:2], collapse = "-")) %>% 
  str_replace("3/4", "3 and 4")

walk2(plts, out_ids, 
    ~save_plot(file.path(gp4_fig_dir, 
                         paste0("umap_gp4_by_", .y, ".pdf")),
          .x, 
          base_asp = 1))

```

### plot various gnes

Most of the tSNE plots show really sparse and low expression of these genes. Not sure how reliable those finding are. 

```{r}

famous_genes <- c(
  "EOMES",
  "MYC",
  "NRL",
  "TOP2A",
  "CRX",
  "ROM1",
  "GRIA2"
)

plts <- plot_harmony(so, 
               famous_genes, 
               minimal_theme = TRUE)
names(plts) <- famous_genes
iwalk(plts, 
    ~save_plot(file.path(gp4_fig_dir, 
                         paste0("umap_gp4_by_", .y, ".pdf")),
          .x, 
          base_asp = 1))
```


### cell count tables 


```{r}
res <- so@meta.data %>% 
  group_by(coarse_clusters_gp4_harmony, UPN) %>% 
  summarize(n = n()) %>% 
  ungroup() %>% 
  spread(UPN, n, fill = 0L) 

write_csv(res, file.path(tbls_dir, "gp4",  "gp4_cells_per_cluster_per_sample.csv"))

res <- so@meta.data %>% 
    group_by(coarse_clusters_gp4_harmony, UPN) %>% 
    summarize(n = n()) %>%
  group_by(coarse_clusters_gp4_harmony) %>% 
  mutate(total = sum(n),
         percent = 100 * n / total) %>% 
  ungroup() %>% 
  select(-n, -total) %>% 
  spread(UPN, percent, fill = 0L) 

  
write_csv(res, file.path(tbls_dir, "gp4",
                  "gp4_percent_cells_per_cluster_per_sample.csv"))

res <- so@meta.data %>%
    group_by(coarse_clusters_gp4_harmony, subtype) %>%
  summarize(n = n()) %>%
  ungroup() %>%
  spread(subtype, n, fill = 0L)


write_csv(res, file.path(tbls_dir, "gp4",
                  "gp4_subtypes_per_cluster.csv"))

res <- so@meta.data %>%
    group_by(coarse_clusters_gp4_harmony, subgroup) %>%
  summarize(n = n()) %>%
  ungroup() %>%
  spread(subgroup, n, fill = 0L)


write_csv(res, file.path(tbls_dir, "gp4",
                  "gp4_subgroups_per_cluster.csv"))
  
```


# Immune subsets


```{r}
so <- qread(file.path("objects", "immune", "immune_so.qs"))

immune_fig_dir <- file.path(fig_dir, "immune")
dir.create(immune_fig_dir, showWarnings = FALSE)
```


### Various

```{r}

p <- plot_umap(so, 
             "fine_immune_clusters", 
             label_color = "black", 
             label_text = TRUE) +
  labs(x = "UMAP 1",
       y = "UMAP 2") +
  theme(legend.position = "none")

save_plot(file.path(immune_fig_dir, "umap_by_clusters.pdf"),
          p, 
          base_asp = 1.0)

p <- plot_umap(so, 
             "UPN", legend_title = "", sorted = "random")

save_plot(file.path(immune_fig_dir, "umap_by_upn.pdf"),
          p, 
          base_asp = 1.4)

p <- plot_umap(so, 
             "subgroup", .cols = palette_OkabeIto, legend_title = "")

save_plot(file.path(immune_fig_dir, "umap_by_subgroup.pdf"),
          p, 
          base_asp = 1.4)


p <- plot_umap(so, "new_cell_type_ids", legend_title = "")

save_plot(file.path(immune_fig_dir, "umap_by_cell_type.pdf"),
          p, 
          base_asp = 2.0)

p <- plot_umap(so, "new_cell_type_ids", group = "subgroup", legend_title = "")

p
save_plot(file.path(immune_fig_dir,
                    "umap_by_fine_cell_types_split_by_subgroup.pdf"),
          p, base_asp = 1.4,
          nrow = 2, 
          ncol = 2)

```


```{r}
p <- plot_cell_proportions(so, "new_cell_type_ids", "subgroup") +
  labs(x = "")

save_plot(file.path(immune_fig_dir, 
                    "barplot_cell_proportions.pdf"),
          p,
          base_asp = 0.75,
          base_height = 6)



plt_dat <- so@meta.data %>% 
  group_by(UPN, subgroup) %>% 
  mutate(total_cells = n()) %>% 
  group_by(UPN, new_cell_type_ids, subgroup) %>% 
  summarize(percent_of_cells = n() / unique(total_cells)) 

 p <- plt_dat %>% 
  ggplot(aes(UPN, percent_of_cells)) +
  geom_col(aes(fill = new_cell_type_ids)) +
  facet_grid(~subgroup, scales = "free_x",  space = "free_x") +
  scale_fill_manual(name = "", 
                    values = discrete_palette_default) +
  labs(x = "",
       y = "Percent of Cells") +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5))

p

save_plot(file.path(immune_fig_dir, 
                    "barplot_cell_percentages_per_tumor.pdf"),
          p,
          base_asp = 2,
          base_height = 6)

```

Need to rerun, splitting out myleoid from lymphocytes

```{r, eval = FALSE}
to_plot <- c(
  "Activated-Microglia",
  "Dendritic-cell-like",
  "Microglia",
  "Neutrophil"
)
so_sub <- subset(so, subset = new_cell_type_ids_short %in% to_plot)

p <- plot_umap(so_sub, "new_cell_type_ids", group = "subgroup",
                         legend_title = "") +
  guides(color = guide_legend(nrow = 3, override.aes = list(size = 6))) + 
  theme(legend.position = "top") + 
  coord_cartesian(xlim = c(-7, 7), ylim = c(-14, 0))

p
save_plot(file.path(immune_fig_dir,
                    "umap_by_myleoid_cell_types_split_by_subgroup.pdf"),
          p, base_asp = 1,
          nrow = 2, 
          ncol = 2)


lym_cell_types <- c(
  "Unknown1",
  "T-cells",
  "NK-cells",
  "Treg"
)

so_sub <- so[, so$new_cell_type_ids_short %in% lym_cell_types]

p <- plot_features_split(so_sub, "new_cell_type_ids", "subgroup",
                         legend_title = "") +
  guides(color = guide_legend(nrow = 2, override.aes = list(size = 6))) + 
  theme(legend.position = "top") + 
  coord_cartesian(xlim = c(-7.5, 2), ylim = c(5.5, 16))

p
save_plot(file.path(immune_fig_dir,
                    "umap_by_t_cell_types_split_by_subgroup.pdf"),
          p, base_asp = 1,
          nrow = 2, 
          ncol = 2)

genes_to_plot <- c(
  "GZMK", #GranzymeK
  "PDCD1", #PD1
  "LAG3",# exhaustion marker
  "IL2",
  "IFNG"
  )
)

```

### cluster markers

```{r fig.height = 6}
mkrs <- read_tsv(file.path(mkrs_dir, "immune", "immune_cell_type_markers.tsv"))

topx <- mkrs %>% 
  arrange(padj) %>% 
  group_by(group) %>% 
  slice(1:10) 

p <- DotPlot(so, 
        features = unique(topx$feature),
        cols = "Spectral",
        group.by = "new_cell_type_ids") +
  coord_flip() +
  labs(y = "",
       x = "") +
  theme(axis.text.x = element_text(size = 24, angle = 90, hjust = 0.5),
        axis.text.y = element_text(size = 10),
        axis.title = element_blank())

p

save_plot(file.path(immune_fig_dir, 
                    "dotplot_immune_markers.pdf"), 
          p, base_height = 24, base_asp = 0.5)
```

## heatmap Myeloid

```{r}
to_plot <- c(
  "Dendritic-cell-like (CD1C+,FCER1A+)",
  "TAM M1/M2 (MRC1+ CD163+)",
  "TAM M1/M2 (APOE+, TYROBP+)",
  "TAM CytoHi+ (CCL3+, SPP1+)",
  "Neutrophil/Monocyte (S100A9+, THBS1+)",
  "Microglia (P2RY12+ TMEM119+)"
)

topx <- read_tsv(file.path(mkrs_dir, "immune" , "immune_cell_type_markers.tsv")) %>% 
  filter(padj < 0.05,
         group %in% to_plot) %>% 
  group_by(group) %>% 
  arrange(padj, desc(logFC), .by_group = TRUE) %>% 
  slice(1:10)

all_ids <- sort(unique(so$new_cell_type_ids)) 
cols_to_use <- discrete_palette_default[which(all_ids %in% to_plot)]

so <- ScaleData(so, features = rownames(so))

cells_to_keep <- colnames(so)[which(so$new_cell_type_ids %in% to_plot)]
tmp_so <- subset(so, cells = cells_to_keep)

tmp_so$new_cell_type_ids <- factor(tmp_so$new_cell_type_ids)

p <- DoHeatmap(tmp_so, 
               group.colors = cols_to_use,
               features = unique(topx$feature),
               group.by = "new_cell_type_ids",
               angle = 45, 
               size = 4,
               raster = FALSE, 
               draw.lines = TRUE)

p <- p +
  scale_fill_gradientn(colours = viridis::viridis(256),
                       name = "Expression\nZ-scores")

save_plot(file.path(immune_fig_dir, "heatmap_markers_microglia.pdf"), 
          p,
          base_width = 10,
          base_height = 12)
```


## heatmap Lymphocytes

```{r}
to_plot <- c(
  "Proliferative (MKI67+, TOP2A+)",
  "Treg (FOXP3+, CTLA4+)",
  "NK-cells (GNLY+, NKG7+)",
  "T-cells (CD3G+, IL32+)",
  "B-cells (CD79A+, IGHM+)"
)

all_ids <- sort(unique(so$new_cell_type_ids)) 
cols_to_use <- discrete_palette_default[which(all_ids %in% to_plot)]

topx <- read_tsv(file.path(mkrs_dir, "immune", "immune_cell_type_markers.tsv")) %>% 
  filter(padj < 0.05,
         group %in% to_plot) %>% 
  group_by(group) %>% 
  arrange(padj, desc(logFC), .by_group = TRUE) %>% 
  slice(1:10)

cells_to_keep <- colnames(so)[which(so$new_cell_type_ids %in% to_plot)]
tmp_so <- subset(so, cells = cells_to_keep)

tmp_so$new_cell_type_ids <- factor(tmp_so$new_cell_type_ids)

p <- DoHeatmap(tmp_so, 
               group.colors = cols_to_use,
               features = unique(topx$feature),
               group.by = "new_cell_type_ids",
               angle = 45, 
               size = 4,
               raster = FALSE, 
               draw.lines = TRUE)

p <- p +
  scale_fill_gradientn(colours = viridis::viridis(256),
                       name = "Expression\nZ-scores")

save_plot(file.path(immune_fig_dir, "heatmap_markers_lymphocytes.pdf"), 
          p,
          base_width = 10,
          base_height = 14)
```

### plot various gnes

#### Myeloid marker phenotypes

Need to redo, with reprojected myeloid or lymphocyte
```{r, eval = FALSE}

myleoid_markers <- list(
  general_markers = c(
  "THBS1", #MDSC or Neutropils
  "P2RY12", #Microglia
  "CD68", # Macrophages
  "C1QA", #TAM
  "TREM2", #TAM
  "APOE" #TAM
),
  m2_m1_markers = c(
  "HLA-DRB1", #M1,
  "CD86", #M1 M2c,
  "TLR2", #M1
  "CD163", #M2
  "MRC1", #M2
  "FCGR3A" #M2
  ),
cytokines = c(
  "CCL3",
  "TNF", #M1
  "CXCL8", #M2 inflammatory (Gr3&4)
  "IL1B", #M2 inflammatory (Grp3&4),
  "CD274", #PDL1
  "PDCD1LG2" #PDL2
  )
)

myl_cell_types <- c(
  "Dendritic-cell-like",
  "Activated-Microglia",
  "Neutrophil",
  "Microglia"
)

myl_so <- so[, so$new_cell_type_ids_short %in% myl_cell_types]
  
umap_lims <- coord_cartesian(xlim = c(-7, 7), ylim = c(-14, 0))
plts <- imap(myleoid_markers,
            function(mkrs, id) {
  n <- length(mkrs)
  n_col <- ceiling(n / 2)
  n_row <- ceiling(n / 3)
  plt <- map(mkrs, 
              ~plot_umap(myl_so, 
                .x, 
                minimal_theme = TRUE) + 
               umap_lims) %>% 
    plot_grid(plotlist= .,
              nrow = n_row,
              ncol = n_col)

  save_plot(file.path(immune_fig_dir, 
                         paste0("umap_", id, "_markers.pdf")),
          plt, 
          nrow = n_row,
          ncol = n_col,
          base_asp = 1)
})


```


#### Lymphocyte marker phenotypes
```{r, eval = FALSE}

lymphocyte_markers <- list(
  general_lymph_markers = c(
  "CD8A", #cd4 t
  "CD4", #cd8 t
  "FOXP3", # Treg
  "KLRD1" #NK cells
),
lymph_cytokines = c(
  "PRF1", #Perforin
  "GZMK", #GranzymeK
  "PDCD1", #PD1
  "LAG3",# exhaustion marker
  "IL2",
  "IFNG"
  )
)

lym_cell_types <- c(
  "Unknown1",
  "T-cells",
  "NK-cells",
  "Treg"
)

lym_so <- so[, so$new_cell_type_ids_short %in% lym_cell_types]
  
umap_lims <- coord_cartesian(xlim = c(-7.5, 2), ylim = c(5.5, 16))
plts <- imap(lymphocyte_markers,
            function(mkrs, id) {
  n <- length(mkrs)
  n_col <- ceiling(n / 2)
  n_row <- ceiling(n / 3)
  plt <- map(mkrs, 
              ~plot_feature(lym_so, 
                .x, 
                embedding = "umap", 
                minimal_theme = TRUE) + 
               umap_lims) %>% 
    plot_grid(plotlist= .,
              nrow = n_row,
              ncol = n_col)

  save_plot(file.path(immune_fig_dir, 
                         paste0("umap_", id, "_markers.pdf")),
          plt, 
          nrow = n_row,
          ncol = n_col,
          base_asp = 1)
})


```




### cell count tables 


```{r}
res <- so@meta.data %>% 
  group_by(new_cell_type_ids, UPN) %>% 
  summarize(n = n()) %>% 
  ungroup() %>% 
  spread(UPN, n, fill = 0L) 

write_csv(res, file.path(tbls_dir, "immune",
                  "immune_cells_per_celltype_per_sample.csv"))

res <- so@meta.data %>% 
    group_by(new_cell_type_ids, UPN) %>% 
    summarize(n = n()) %>%
  group_by(new_cell_type_ids) %>% 
  mutate(total = sum(n),
         percent = 100 * n / total) %>% 
  ungroup() %>% 
  select(-n, -total) %>% 
  spread(UPN, percent, fill = 0L) 

  
write_csv(res, file.path(tbls_dir, "immune",
                  "immune_cells_percent_cells_per_celltype_per_sample.csv"))

res <- so@meta.data %>%
    group_by(new_cell_type_ids, subgroup) %>%
  summarize(n = n()) %>%
  ungroup() %>%
  spread(subgroup, n, fill = 0L)

res

write_csv(res, file.path(tbls_dir, "immune",
                  "immune_subgroups_per_cell_type.csv"))
  
```

# Combined tumors

```{r}
so_fns <- c(
  GP3 = file.path("objects", "gp34", "gp3_only.qs"),
  GP4 = file.path("objects", "gp34", "gp4_only.qs"),
  SHH = file.path("objects", "shh", "shh.qs")
)

sos <- map(so_fns, qread)

clusters <- c("tumor_subpopulations_gp3", 
              "tumor_subpopulations_gp4", 
              "tumor_subpopulations_shh")

sos <- map2(sos, clusters, ~{
  to_keep <- colnames(.x)[!str_detect(.x@meta.data[[.y]], "X")]
  subset(.x, cells = to_keep)
})


```

```{r}
mkr_list <- list(
  GP3 = c("GABRA5",
          "IMPG2",
          "EYS",
          "NRL",
          "MAB21L2",
          "NPR3"),
  GP4 = c("RBM24",
          "UNC5D",
          "KCNA1",
          "EOMES",
          "KHDRBS2",
          "OAS1"),
  SHH = c("PDLIM3",
          "EYA1",
          "HHIP",
          "ATOH1",
          "SFRP1")
)

clusters <- c("tumor_subpopulations_gp3", 
              "tumor_subpopulations_gp4", 
              "tumor_subpopulations_shh")

plts <- imap(mkr_list,
    function(mkrs, id) {
      all_subgroups_plot <- list()
      for(i in seq_along(sos)){
        cluster_col <- clusters[i]
        so <- sos[[i]]
        plts <- list()
        to_plot <- c(cluster_col, mkrs)
        for(j in seq_along(to_plot)){

          if(j == 1){
            plts[[j]] <-  plot_harmony(so, cluster_col, 
                    legend_title = "", 
                    minimal_theme = TRUE,
                    label_text = TRUE,
                    label_color = "black",
                    ggrepel_opts = list(force = 0)) +
              labs(title = "")
          } else {
            plts[[j]] <- plot_harmony(so, to_plot[j], 
                    legend_title = "", 
                    minimal_theme = TRUE,
                    label_text = TRUE,
                    label_color = "black",
                    ggrepel_opts = list(force = 0)) +
              labs(title = ifelse(i == 1, to_plot[j], "")) +
              theme(title = element_text(size = 30))
          }
        }
        all_subgroups_plot[[i]] <- plot_grid(plotlist = plts, nrow = 1, ncol = length(mkrs) + 1)
        
      }
      p <- plot_grid(plotlist = all_subgroups_plot, 
                nrow = 3, ncol = 1)
      
      save_plot(file.path(all_neo_fig_dir, 
                           str_c(id, "_marker_overview.pdf")),
                 p, 
                 nrow = 3, 
                 ncol = 7, 
                 base_asp = 1)
      p
})



```
